### 2019-12-13 #1955 @whedon
[PRE REVIEW]: Soss: Code Generation for Probabilistic Programming in Julia
**Submitting author:** @cscherrer (<a href="http://orcid.org/0000-0002-1490-0304">Chad Scherrer</a>)
**Repository:** <a href="https://github.com/cscherrer/Soss.jl" target ="_blank">https://github.com/cscherrer/Soss.jl</a>
**Version:** v0.9.0
