### 2019-12-13 #1955 @cscherrer
Hi @kyleniemeyer , I just saw the "if you have any suggestions for potential reviewers...". both Matt Sottile and Dan Foreman-Mackey are familiar with this work (though neither are directly involved, so no conflict of interest). Outside existing JOSS reviewers, any contributor to [Turing.jl](https://github.com/TuringLang/Turing.jl) or [Gen](https://github.com/probcomp/Gen) could be a good fit.
