### 2019-12-14 5217706
get symlogpdf(::Poisson) working
