### 2019-12-17 db70731
logfactorial
