+++ src/Soss.jl
@@ -38,6 +38,7 @@ include("primitives/likelihood-weighting.jl")
 
 include("transforms/predictive.jl")
 include("transforms/markovblanket.jl")
+include("transforms/do.jl")
 
 include("symbolic/symbolic.jl")
 include("symbolic/codegen.jl")
