### 2019-12-31 478f374
particles(::Vector{Array{T,N}})
