+++ src/core/utils.jl
@@ -126,7 +126,7 @@ function buildSource(m, proc, wrap=identity; kwargs...)
 
     kernel = @q begin end
 
-    for st in map(v -> findStatement(m,v), toposortvars(m))
+    for st in map(v -> findStatement(m,v), toposort(m))
         ex = proc(m, st; kwargs...)
         isnothing(ex) || push!(kernel.args, ex)
     end
