+++ src/core/utils.jl
@@ -269,7 +269,7 @@ function tower(x)
     return result
 end
 
-TypeLevel = GeneralizedGenerated.TypeLevel
+const TypeLevel = GeneralizedGenerated.TypeLevel
 
 unVal(::Type{Val{T}}) where {T} = T
 unVal(::Val{T}) where {T} = T
