+++ README.md
@@ -277,4 +277,3 @@ We need ways to interact with Turing and Gen. Some ideas:
 
 [![Stargazers over time](https://starchart.cc/cscherrer/Soss.jl.svg)](https://starchart.cc/cscherrer/Soss.jl)
 
-
