+++ Project.toml
@@ -49,7 +49,7 @@ LazyArrays = "0.14"
 LogDensityProblems = "0.9"
 MLStyle = "0.3"
 MacroTools = "0.5"
-MonteCarloMeasurements = "0.5"
+MonteCarloMeasurements = "0.5,0.6"
 NamedTupleTools = "0.12"
 Plots = "0.28"
 PyCall = "1.91"
