### 2020-01-10 #1955 @melissawm
@cscherrer sorry for taking so long to respond, the holidays were a busy period for me. I know you indicated some names above, but I can't find them in the list of past JOSS reviewers. I'll try some people from our reviewers list first, if that's ok.
