### 2020-01-26 9ea29c1
add predict(m::Model; kwargs...)
