### 2020-01-27 #1955 @melissawm
@kose-y Hello, would you be willing/available to review this software paper for JOSS?
