### 2020-02-05 #2062 @kose-y
Hi @cscherrer, the link to the document on the badge on your repository does not seem to work.

https://cscherrer.github.io/Soss.jl/stable
https://cscherrer.github.io/Soss.jl/dev

Is there another way to view the web version of documentation?
