### 2020-02-06 #2062 @kose-y
SymPy is listed twice in the references of the paper.
