### 2020-02-06 #2062 @cscherrer
Thanks @kose-y !

Documentation is currently limited to the README. I do have some slides and a PROBPROG submission I could point to, but generated docs aren't yet set up.

Thanks for catching the SymPy double-citation, I'll get that fixed. Feel free to add issues directly in the repo if you like, or I can add them as we go :)
