### 2020-02-12 540f731
docs: adds installation instructions to the readme
