+++ Project.toml
@@ -51,7 +51,7 @@ MLStyle = "0.3"
 MacroTools = "0.5"
 MonteCarloMeasurements = "0.6,0.7"
 NamedTupleTools = "0.12"
-Plots = "0.28"
+Plots = "0.28, 0.29"
 PyCall = "1.91"
 Reexport = "0.2"
 ResumableFunctions = "0.5"
