+++ src/inference/advancedhmc.jl
@@ -12,7 +12,7 @@ using AdvancedHMC
 export advancedHMC
 
 
-function advancedHMC(m :: JointDistribution{A,B}, _data, n_samples = 1000; 
+function advancedHMC(m :: JointDistribution{A,B}, _data, N = 1000; 
                                                          n_adapts  = 1000) where {A,B}
     
     ℓ(pars) = logpdf(m, merge(pars, _data))
@@ -47,7 +47,7 @@ function advancedHMC(m :: JointDistribution{A,B}, _data, n_samples = 1000;
     # Draw samples via simulating Hamiltonian dynamics
     # - `samples` will store the samples
     # - `stats` will store statistics for each sample
-    samples, stats = sample(h, prop, x_init, n_samples, adaptor, n_adapts; progress=true)
+    samples, stats = sample(h, prop, x_init, N, adaptor, n_adapts; progress=true)
 
 end
 
