### 2020-02-13 cec5731
CompatHelper: bump compat for "LazyArrays" to "0.15" (#113)
Co-authored-by: Seth Axen <seth.axen@gmail.com>
Co-authored-by: Chad Scherrer <chad.scherrer@gmail.com>
