+++ Project.toml
@@ -45,7 +45,7 @@ ForwardDiff = "0.10"
 GeneralizedGenerated = "0.2"
 Graphs = "0.10"
 IterTools = "1.3"
-LazyArrays = "0.14"
+LazyArrays = "0.14, 0.15"
 LogDensityProblems = "0.10"
 MLStyle = "0.3"
 MacroTools = "0.5"
