+++ Project.toml
@@ -19,10 +19,10 @@ MLStyle = "d8e11817-5142-5d16-987a-aa16d5891078"
 MacroTools = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
 MonteCarloMeasurements = "0987c9cc-fe09-11e8-30f0-b96dd679fdca"
 NamedTupleTools = "d9ec5142-1e00-5aa0-9d6a-321866360f50"
-Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 PyCall = "438e738f-606a-5dbb-bf0a-cddfbfd45ab0"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+RecipesBase = "3cdcf5f2-1ef4-517c-9805-6587b60abb01"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 ResumableFunctions = "c5292f4c-5179-55e1-98c5-05642aab7184"
 ReverseDiff = "37e2e3b7-166d-5795-8a7a-e32c996b4267"
@@ -51,8 +51,8 @@ MLStyle = "0.3"
 MacroTools = "0.5"
 MonteCarloMeasurements = "0.6,0.7"
 NamedTupleTools = "0.12"
-Plots = "0.28, 0.29"
 PyCall = "1.91"
+RecipesBase = "0.7,0.8"
 Reexport = "0.2"
 ResumableFunctions = "0.5"
 ReverseDiff = "1.0"
@@ -66,8 +66,9 @@ julia = "^1.3"
 
 [extras]
 BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 Weave = "44d3d7a6-8a23-5bf8-98c5-b353f8df5ec9"
 
 [targets]
-test = ["Test", "Weave", "BenchmarkTools"]
+test = ["Plots", "Test", "Weave", "BenchmarkTools"]
