### 2020-02-14 0da5af5
CompatHelper: bump compat for "MonteCarloMeasurements" to "0.8"
