+++ Project.toml
@@ -49,7 +49,7 @@ LazyArrays = "0.14, 0.15"
 LogDensityProblems = "0.10"
 MLStyle = "0.3"
 MacroTools = "0.5"
-MonteCarloMeasurements = "0.6,0.7"
+MonteCarloMeasurements = "0.6,0.7, 0.8"
 NamedTupleTools = "0.12"
 PyCall = "1.91"
 RecipesBase = "0.7,0.8"
