### 2020-02-14 c493548
docs: move installation istructions to README.jmd
