### 2020-02-14 dc80e0d
docs: adds installation instructions to the readme (#104)
* docs: adds installation instructions to the readme

* Mark code blocks as julia

Co-Authored-By: Seth Axen <seth.axen@gmail.com>

* docs: move installation istructions to README.jmd

Co-authored-by: Seth Axen <seth.axen@gmail.com>
