+++ README.md
@@ -7,12 +7,36 @@
 [![Codecov](https://codecov.io/gh/cscherrer/Soss.jl/branch/master/graph/badge.svg)](https://codecov.io/gh/cscherrer/Soss.jl)
 [![Coveralls](https://coveralls.io/repos/github/cscherrer/Soss.jl/badge.svg?branch=master)](https://coveralls.io/github/cscherrer/Soss.jl?branch=master)
 
-Soss is a library for _probabilistic programming_
+Soss is a library for _probabilistic programming_.
+
+## Getting started
+
+Soss is an officially registered package, so to add it to your project you can type
+````julia
+
+]add Soss
+````
+
+
+
+within the julia REPL and your are ready for `using Soss`. If it fails to precompile, it could be due to one of the following:
+
+* You have gotten an old version due to compatibility restrictions with your current environment.
+Should that happen, create a new folder for your Soss project, launch a julia session within, type
+````julia
+
+]activate .
+````
+
+
+
+and start again. More information on julia projects [here](https://julialang.github.io/Pkg.jl/stable/environments/#Creating-your-own-projects-1).
+* You have set up PyCall to use a python distribution provided by yourself. If that is the case, make sure to install the missing python dependencies, as listed in the precompilation error. More information on PyCall's python version [here](https://github.com/JuliaPy/PyCall.jl#specifying-the-python-version).
 
 Let's jump right in with a simple linear model:
 
 ````julia
-using Soss 
+using Soss
 
 m = @model X begin
     β ~ Normal() |> iid(size(X,2))
@@ -91,14 +115,14 @@ julia> truth.y
 
 
 
-And now pretend we don't know `β`, and have the model figure it out. 
+And now pretend we don't know `β`, and have the model figure it out.
 Often these are easier to work with in terms of `particles` (built using [MonteCarloMeasurements.jl](https://github.com/baggepinnen/MonteCarloMeasurements.jl)):
 
 ````julia
 julia> post = dynamicHMC(m(X=truth.X), (y=truth.y,));
 
 julia> particles(post)
-(β = Particles{Float64,1000}[0.548 ± 0.24, 0.751 ± 0.5],)
+(β = Particles{Float64,1000}[0.538 ± 0.26, 0.775 ± 0.51],)
 
 ````
 
@@ -132,12 +156,12 @@ truth.y - particles(ppc)
 
 
 ````
-6-element Array{MonteCarloMeasurements.Particles{Float64,1000},1}:
- -0.52 ± 0.55 
- -1.21 ± 1.3  
-  0.57 ± 0.53 
-  0.951 ± 0.91
-  0.655 ± 0.63
+6-element Array{Particles{Float64,1000},1}:
+ -0.534 ± 0.55
+ -1.28 ± 1.3  
+  0.551 ± 0.53
+  0.918 ± 0.91
+  0.624 ± 0.63
   0.534 ± 0.53
 ````
 
@@ -161,16 +185,23 @@ end;
 
 julia> 
 symlogpdf(m2)
-   N                                                     k                                                         
-  ____                                                  ____                                                       
-  ╲                                                     ╲                                                          
-   ╲    ⎛                      2                  ⎞      ╲    ⎛            2                       log(π)   log(2)⎞
-    ╲   ⎜  (y[_j1] - yhat[_j1])    log(π)   log(2)⎟       ╲   ⎜- 0.5⋅β[_j1]  - 0.693147180559945 - ────── + ──────⎟
-    ╱   ⎜- ───────────────────── - ────── - ──────⎟ +     ╱   ⎝                                      2        2   ⎠
-   ╱    ⎝            2               2        2   ⎠      ╱                                                         
-  ╱                                                     ╱                                                          
-  ‾‾‾‾                                                  ‾‾‾‾                                                       
[…]
