### 2020-02-15 #2062 @kose-y
I have checked and read documentations, your outside slides, and PROBPROG submission. I think this is a fantastic package resolving the "three-language problem" often arising from Bayesian inference (Although I'm not a Bayesian, I have used JAGS in a course before, and have seen how TensorFlow Probability and PyMC3 works.). However, the review process of JOSS is checklist-driven, and I have to comment based on the checklist above.

- On a statement of need: I think the "three-language problem" claim deserves to be included in the software paper as well as `README.md`.
- Functionality documentation: current information given on `README.md` is weak. For example, I could not figure out what `codegen` means from `README.md` alone. I think including the description from the PROBPROG submission would help.
- Automated tests: All I could find was a test checking if `README.md` is successfully generated from `README.jmd`. I think it should check if each functionality works correctly.
- Community guidelines: It is required in the documentation.
