### 2020-02-15 955dba7
Switch to new `test/Project.toml` approach
