+++ Project.toml
@@ -63,12 +63,3 @@ StatsFuns = "0.9"
 SymPy = "1.0"
 TransformVariables = "0.3"
 julia = "^1.3"
-
-[extras]
-BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
-Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-Weave = "44d3d7a6-8a23-5bf8-98c5-b353f8df5ec9"
-
-[targets]
-test = ["Plots", "Test", "Weave", "BenchmarkTools"]
