### 2020-02-15 56711f5
update Project.toml and .travis.yml
