+++ docs/Project.toml
@@ -1,2 +1,3 @@
 [deps]
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+Soss = "8ce77f84-9b61-11e8-39ff-d17a774bf41c"
