### 2020-02-16 a051ab7
CompatHelper: add new compat entry for "SpecialFunctions" at version "0.10" (#110)
Co-authored-by: Seth Axen <seth.axen@gmail.com>
