+++ Project.toml
@@ -59,6 +59,7 @@ ReverseDiff = "1.0"
 SimpleGraphs = "0.3"
 SimplePartitions = "0.2"
 SimplePosets = "0.0"
+SpecialFunctions = "0.10"
 StatsFuns = "0.9"
 SymPy = "1.0"
 TransformVariables = "0.3"
