### 2020-02-16 999602f
Regex magic from Seth
