### 2020-02-16 f04ffae
Docs (#118)
* Switch to new `test/Project.toml` approach

* update Project.toml and .travis.yml

* Make docs work (maybe?)

* Regex magic from Seth

* 404 fixin
