### 2020-02-18 cscherrer/Soss.jl#119 @kose-y
JOSS Review
https://github.com/openjournals/joss-reviews/issues/2062#issuecomment-586559646
