### 2020-02-20 cbc3853
feat: towards GPU support
