+++ src/distributions/for.jl
@@ -3,40 +3,56 @@ import Distributions.logpdf
 using Base.Cartesian
 using Base.Threads
 using FillArrays
+using CuArrays
+
+include("gpulogpdfs.jl")
 
 export logpdf
 export rand
 
 export For
-struct For{F,T,D,X} 
-    f :: F  
-    θ :: T
+struct For{D,P,X,N}
+    params :: Array{P,N}
 end
 
 #########################################################
 # T <: NTuple{N,J} where {J <: Integer}
 #########################################################
 
-For(f, θ::J...) where {J <: Integer} = For(f,θ)
+For(f, dims::J...) where {J <: Integer} = For(f, dims)
 
-function For(f::F, θ::T) where {F, N, J <: Integer, T <: NTuple{N,J}}
-    d = f.(Ones{Int}(N)...)
+function For(f, dims::T) where {N, J<:Integer, T<:NTuple{N,J}}
+    θ = one(CartesianIndex{N})
+    d = f(Tuple(θ)...)
     D = typeof(d)
     X = eltype(d)
-    For{F, NTuple{N,J}, D, X}(f,θ)
+    p = params(d)
+    P = typeof(p)
+    ps = Array{P,N}(undef, dims)
+    ps[θ] = p
+    @inbounds @simd for θ in CartesianIndices(dims)[2:end]
+        d = f(Tuple(θ)...)
+        ps[θ] = params(d)
+    end
+    For{D,P,X,N}(ps)
 end
 
-@inline function logpdf(d::For{F,T,D,X1},xs::AbstractArray{X2,N}) where {F, N, J <: Integer, T <: NTuple{N,J}, D,  X1,  X2}
+@inline function logpdf(d::For{D}, xs::AbstractArray) where D
     s = 0.0
-    @inbounds @simd for θ in CartesianIndices(d.θ)
-        s += logpdf(d.f(Tuple(θ)...), xs[θ])
+    for (p, x) in zip(d.params, xs)
+        s += logpdf(D(p...), x)
     end
     s
 end
 
-function Base.rand(dist::For) 
-    map(CartesianIndices(dist.θ)) do I
-        (rand ∘ dist.f)(Tuple(I)...)
+@inline function logpdf(d  :: For{Normal{X}, Tuple{X,X}, X, N},
+                        xs :: CuArray{<:Real, N}) where {X<:Real, N}
+    sum(cunormlogpdf.(d.params |> cu, xs))
+end
+
+function Base.rand(dist::For{D}) where D
+    map(dist.params) do p
+        (rand ∘ D)(p...)
     end
 end
 
@@ -44,102 +60,94 @@ end
 # T <: NTuple{N,J} where {J <: AbstractUnitRange}
 #########################################################
 
-For(f, θ::J...) where {J <: AbstractUnitRange} = For(f,θ)
-
-function For(f::F, θ::T) where {F, N, J <: AbstractRange, T <: NTuple{N,J}}
-    d = f.(ones(Int, N)...)
-    D = typeof(d)
-    X = eltype(d)
-    For{F, NTuple{N,J}, D, X}(f,θ)
-end
-
-
-@inline function logpdf(d::For{F,T,D,X1},xs::AbstractArray{X2,N}) where {F, N, J <: AbstractRange,  T <: NTuple{N,J}, D, X1, X2}
-    s = 0.0
-    @inbounds @simd for θ in CartesianIndices(d.θ)
-        s += logpdf(d.f(Tuple(θ)...), xs[θ])
-    end
-    s
-end
-
-
-function Base.rand(dist::For{F,T}) where {F,  N, J <: AbstractRange, T <: NTuple{N,J}}
[…]
