### 2020-02-20 4d10c82
feat: adds Inf warning when computing logpdf
