### 2020-02-22 c905a7c
feat: towards supporting logpdf calculations on gpu
* For has been heavily modified (work in progress). Now it computes 
parameters on construction. Objective: compute parameters directly on 
GPU.

* iid has been modified. Some iid logpdf calcuations already work on 
GPU. This has required modifications in canonical, too.

* CuArrays logpdfs for Uniform, Normal and Gamma distributions (work in 
progress, refactoring is needed).
