+++ src/distributions/gpulogpdfs.jl
@@ -1,34 +1,25 @@
 using CUDAnative
-# using Distributions
-# using Distributions: zval, log2π
+using Distributions: log2π
+using CuArrays: @cufunc
+import StatsFuns: normlogpdf, gammalogpdf
 
-Base.@irrational log2π  1.8378770664093454836 log(big(2.)*π)
-Base.@irrational sqrt2π 2.5066282746310005024 sqrt(big(π)*2.0)
+# Base.@irrational log2π  1.8378770664093454836 log(big(2.)*π)
+# Base.@irrational sqrt2π 2.5066282746310005024 sqrt(big(π)*2.0)
 # Base.@irrational invsqrt2π 0.3989422804014326779 inv(big(sqrt2π))
 
 
+# Uniform:
+unilogpdf(p::Tuple{T,T}, x::T) where T<:Real =
+    p[1] ≤ x ≤ p[2] ? -log(p[2] - p[1]) : -T(Inf)
+@cufunc unilogpdf(p::Tuple{Float64,Float64}, x::Float64) =
+    ifelse(p[1] ≤ x ≤ p[2], -CUDAnative.log(p[2] - p[1]), -Inf)
+@cufunc unilogpdf(p::Tuple{Float32,Float32}, x::Float32) =
+    ifelse(p[1] ≤ x ≤ p[2], -CUDAnative.log(p[2] - p[1]), -Float32(Inf))
+
+# Normal:
 zval(μ::Real, σ::Real, x::Number) = (x - μ) / σ
-
-# # pdf
-# cunormpdf(z::Number) = CUDAnative.exp(-CUDAnative.abs2(z)/2) * invsqrt2π
-# function cunormpdf(μ::Real, σ::Real, x::Number)
-#     if iszero(σ)
-#         if x == μ
-#             z = zval(μ, one(σ), x)
-#         else
-#             z = zval(μ, σ, x)
-#             σ = one(σ)
-#         end
-#     else
-#         z = zval(μ, σ, x)
-#     end
-#     cunormpdf(z) / σ
-# end
-
-# logpdf
-cunormlogpdf(x::Real) = -(x^2 + log2π)/2
-
-function cunormlogpdf(p::Tuple{Real,Real}, x::Real)
+normlogpdf(p::Tuple{Real,Real}, x::Real) = normlogpdf(p[1], p[2], x)
+function _cunormlogpdf(p::Tuple{Real,Real}, x::Real)
     μ, σ = p
     if iszero(σ)
         if x == μ
@@ -40,5 +31,33 @@ function cunormlogpdf(p::Tuple{Real,Real}, x::Real)
     else
         z = zval(μ, σ, x)
     end
-    cunormlogpdf(z) - CUDAnative.log(σ)
+    -(z^2 + log2π)/2 - CUDAnative.log(σ)
+end
+@cufunc normlogpdf(p::Tuple{Real,Real}, x::Real) =
+    _cunormlogpdf(p::Tuple{Real,Real}, x::Real)
+
+
+# Gamma:
+gammalogpdf(p::Tuple{Real,Real}, x::Real) = gammalogpdf(p[1], p[2], x)
+_cugammalogpdf(p::Tuple{Real,Real}, x::Real) =
+    -CUDAnative.lgamma(p[1]) - p[1] * CUDAnative.log(p[2]) + (p[1] - 1) * CUDAnative.log(x) - x / p[2]
+@cufunc gammalogpdf(p::Tuple{Real,Real}, x::Real) =
+    _cugammalogpdf(p::Tuple{Real,Real}, x::Real)
+
+
+###############
+# General stuff, refactor:
+
[…]
