### 2020-02-24 ee6b798
feat: some more GPU logpdf calculations
* For and iid reimplemented as instances of Product

* Still problems related to CuArray broadcasting of GeneralizedGenerated 
functions
