+++ src/distributions/for.jl
@@ -1,188 +1,0 @@
-using Distributions
-import Distributions.logpdf
-using Base.Cartesian
-using Base.Threads
-using FillArrays
-using CuArrays
-# using LoopVectorization
-
-include("gpulogpdfs.jl")
-
-export logpdf
-export rand
-export For
-
-
-# struct For{X, N, D<:Sampleable{X}, P} <: Sampleable{AbstractArray{X, N}} # one would like to write this, but Distributions does not support such generality
-struct For{X, N, D, P<:AbstractArray{T,N} where T}
-    params :: P
-end
-
-Base.eltype(d::For{X,N,D}) where {X,N,D} = D
-Distributions.params(d::For) = d.params
-
-# This is general, refactor:
-@inline function logpdf(d, xs::AbstractArray)
-    sum(logpdf(eltype(d)).(Array(params(d)), xs))
-end
-
-@inline function logpdf(d, xs::CuArray)
-    sum(logpdf(eltype(d)).(CuArray(params(d)), xs))
-end
-
-function Base.rand(d::For)
-    map(d.params) do θ
-        (rand ∘ eltype(d))(θ...)
-    end
-end
-
-
-#########################################################
-# T <: NTuple{N,J} where {J <: Integer}
-#########################################################
-
-For(f, dims::J...) where {J <: Integer} = For(f, dims)
-
-function For(f, dims::T) where {N, J<:Integer, T<:NTuple{N,J}}
-    α = one(CartesianIndex{N})
-    d = f(Tuple(α)...)
-    θ = params(d)
-    θs = Array{typeof(θ),N}(undef, dims)
-    θs[α] = θ
-    @inbounds @simd for α in CartesianIndices(dims)[2:end] # Tried to use @avx here, but there are problems with GeneralizedGenerated. Should sometime understand and perhaps solve them.
-        d = f(Tuple(α)...)
-        θs[α] = params(d)
-    end
-    For{eltype(d),N,typeof(d),typeof(θs)}(θs)
-end
-
-
-#########################################################
-# T <: NTuple{N,J} where {J <: AbstractUnitRange}
-#########################################################
-
-For(f, ranges::J...) where {J <: AbstractUnitRange} = For(f, ranges)
-
-function For(f, ranges::T) where {N, J<:AbstractRange, T<:NTuple{N,J}}
-    αs = CartesianIndices(ranges)
-    α = αs[1]
-    α₀ = α - one(CartesianIndex{N})
-    d = f(Tuple(α)...)
-    θ = params(d)
-    θs = Array{typeof(θ),N}(undef, size(αs))
-    θs[α - α₀] = θ
-    @inbounds @simd for α in αs[2:end]
-        d = f(Tuple(α)...)
-        θs[α - α₀] = params(d)
-    end
-    For{eltype(d),N,typeof(d),typeof(θs)}(θs)
-end
-
-
-#########################################################
-# T <: NTuple{N,J} where {J <: AbstractArray}
-#########################################################
-
-### From Michael Abbott's NamedPlus, which as things stand cannot be added because of compatibility restrictions:
-const newaxis = [CartesianIndex()]
-outer(xs::AbstractArray...) = outer(*, xs...)
-function outer(f::Function, x::AbstractArray, ys::AbstractArray...)
-    dims = ndims(x)
-    views = map(ys) do y
-        newaxes = ntuple(_->newaxis, dims)
-        colons = ntuple(_->(:), ndims(y))
-        view(y, newaxes..., colons...)
-    end
-    Broadcast.broadcast(f, x, views...)
-end
-###
-
[…]
