+++ src/distributions/product.jl
@@ -1,0 +1,187 @@
+using Distributions
+import Distributions.logpdf
+using Base.Cartesian
+using Base.Threads
+using FillArrays
+using CuArrays
+# using LoopVectorization
+
+include("gpulogpdfs.jl")
+
+# export logpdf
+# export rand
+export Product, For, iid
+
+
+# Product of measures of type D, with parameters f.(params...)::AbstractArray{T,N} where T is the type of the parameters of D. Both For and iid will be particular cases.
+# struct Product{D<:Sampleable{X} where X, P} <: Sampleable{AbstractArray{X,N}} # one would like to write this, but Distributions does not support such generality
+struct Product{D, F, P} # P will be a parameter grid (NTuple of views)
+    f :: F # cannot restrict this to be a `Function`, because `GeneralizedGenerated.Closure`s are not `Function`s. Also, it will be convenient to allow for `Nothing` here.
+    params :: P
+    Product{D}(f, params) where D = new{D, typeof(f), typeof(params)}(f, params)
+end
+
+Product{D}(f, params::Union{AbstractRange,AbstractArray}...) where D =
+    Product{D}(f, params)
+
+Product{D}(params::Union{AbstractRange,AbstractArray}...) where D =
+    Product{D}(nothing, params)
+
+Base.eltype(p::Product{D}) where D = D
+# corregir, ahora que P es NTuple:
+# # need to know how many arguments p.f (p::Product) takes:
+# Base.length(::Type{NTuple{N,T}}) where {N,T} = N
+# Base.length(p::Product) = length(p.params)
+# Base.ndims(p::Product) = ndims(p.params)
+# function Base.iterate(d::Product{D,P}, i=1) where {D,P<:AbstractArray}
+#     i-1 < length(d.params) ? (D(d.f(d.params[i]...)...), i+1) : nothing
+# end
+# function Base.iterate(d::Product{D,P}, args...) where {D,P<:ProductIterator}
+
+
+Distributions.params(d::Product) = d.params
+
+@inline function logpdf(p::Product, xs::AbstractArray) # D<:Sampleable{X}
+    sum(logpdf(eltype(p)).(p.f.(p.params...), xs))
+end
+
+@inline function logpdf(p::Product{D,Nothing}, xs::AbstractArray) where D
+    sum(logpdf(D).(p.params..., xs))
+end
+
+function Base.rand(p::Product{D}) where D
+    map(p.f.(p.params...)) do θ
+        (rand ∘ D)(θ...)
+    end
+end
+
+
+##############
+# utils
+##############
+
+### Based on Michael Abbott's `outer` in NamedPlus, which as things stand cannot be added because of compatibility restrictions:
+const newaxis = [CartesianIndex()]
+
+function ngrid(xs...)
+    dims = 0
+    views = map(xs) do x
+        newaxes = ntuple(_->newaxis, dims)
+        colons = ntuple(_->(:), ndims(x))
+        dims += ndims(x)
+        view(x, newaxes..., colons...)
+    end
+end
+
+outer(f::Function, xs::AbstractArray...) = Broadcast.broadcast(f, ngrid(xs...)...)
+
+Base.:(*)(s::Tuple, t::Tuple) = (s..., t...)
+
+Core.nothing(x...) = x
+
+
+#############
+# For
+#############
+# using MLStyle
+#
+# macro deconstruct(f::Expr)
+#     @match f begin
+#         Expr(:->, η, Expr(:block, _, Expr(:call, d, θ...))) =>
[…]

[…]
