### 2020-02-24 f5c1129
fix: Product construction and logpdf methods
