### 2020-02-24 95515ca
feat: handles Categorical
