### 2020-02-24 897b1ed
fix: @cufunc use
