### 2020-03-06 666fc67
CompatHelper: bump compat for "NamedTupleTools" to "0.13"
