+++ Project.toml
@@ -50,7 +50,7 @@ LogDensityProblems = "0.10"
 MLStyle = "0.3"
 MacroTools = "0.5"
 MonteCarloMeasurements = "0.6,0.7, 0.8"
-NamedTupleTools = "0.12"
+NamedTupleTools = "0.12, 0.13"
 PyCall = "1.91"
 RecipesBase = "0.7,0.8"
 Reexport = "0.2"
