### 2020-03-12 #2062 @cscherrer
Thanks for bumping this @danielskatz !

Hi @kose-y, thank you for your helpful comments; I'm sorry to not have responded sooner. This project is my first experience with both Julia's documentation and test system, so the software engineering side of this is taking me some time. The [docs|dev](https://cscherrer.github.io/Soss.jl/dev) button now works, but there's nothing really there yet. I should have some time to populate this somewhat this weekend.

I agree there also need to be more tests; this will be an ongoing effort.
