### 2020-03-15 ee0784f
Generalize `Mix`
