### 2020-03-31 7f48cea
disable sympy for faster loading
