+++ Project.toml
@@ -32,7 +32,6 @@ SimplePosets = "b2aef97b-4721-5af9-b440-0bad754dc5ba"
 SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 StatsFuns = "4c63d2b9-4356-54db-8cca-17b64c39e42c"
-SymPy = "24249f21-da20-56a4-8eb1-6a02cf4ae2e6"
 TransformVariables = "84d833dd-6860-57f9-a1a7-6da5db126cff"
 
 [compat]
@@ -61,6 +60,5 @@ SimplePartitions = "0.2"
 SimplePosets = "0.0"
 SpecialFunctions = "0.10"
 StatsFuns = "0.9"
-SymPy = "1.0"
 TransformVariables = "0.3"
 julia = "^1.3"
