### 2020-03-31 33e1b68
SymPy module
