+++ src/_SymPy.jl
@@ -1,0 +1,4 @@
+module SymPy
+    include("symbolic/symbolic.jl")
+    include("symbolic/codegen.jl")
+end
