### 2020-03-31 df6fede
rename SymPy module "sympy"
