+++ src/_sympy.jl
@@ -1,4 +1,4 @@
-module SymPy
+module sympy
     include("symbolic/symbolic.jl")
     include("symbolic/codegen.jl")
 end
