### 2020-03-31 2afae4d
rename module "SymbolicSymPy"
