+++ src/_sympy.jl
@@ -1,4 +1,4 @@
-module sympy
+module SymbolicSymPy
     include("symbolic/symbolic.jl")
     include("symbolic/codegen.jl")
 end
