### 2020-04-01 4e34e1e
add SymPy to docs and test
