### 2020-04-01 e24ac65
Add SymPy back in (miunderstood Requires.jl approach)
