### 2020-04-01 00aeba2
add SymPy to `[compat]`
