+++ Project.toml
@@ -62,5 +62,6 @@ SimplePartitions = "0.2"
 SimplePosets = "0.0"
 SpecialFunctions = "0.10"
 StatsFuns = "0.9"
+SymPy = "1.0"
 TransformVariables = "0.3"
 julia = "^1.3"
