+++ Project.toml
@@ -55,6 +55,7 @@ NamedTupleTools = "0.12, 0.13"
 PyCall = "1.91"
 RecipesBase = "0.7,0.8"
 Reexport = "0.2"
+Requires = "0.5,1.0"
 ResumableFunctions = "0.5"
 ReverseDiff = "1.0"
 SimpleGraphs = "0.3"
