+++ src/Soss.jl
@@ -41,16 +41,12 @@ include("transforms/predictive.jl")
 include("transforms/markovblanket.jl")
 include("transforms/do.jl")
 
-# include("symbolic/symbolic.jl")
-# include("symbolic/codegen.jl")
-
 include("particles.jl")
 include("plots.jl")
 
 include("inference/rejection.jl")
 include("inference/dynamicHMC.jl")
 include("inference/advancedhmc.jl")
-
 
 # include("weighted.jl")
 #
