+++ src/symbolic/symbolic.jl
@@ -1,377 +1,0 @@
-using MacroTools: @q
-using GeneralizedGenerated
-import PyCall
-using MLStyle
-
-import SymPy
-using SymPy: Sym, symbols, free_symbols
-import SpecialFunctions
-using SpecialFunctions: logfactorial
-
-"""As type encoding a PyObject is unsafe without some hard
-works with reference counting, we simply use a Julia proxy
-to imitate the `sympy` module, e.g.,
-    `sympy.attr = _pysympy.attr`
-"""
-struct PySymPyModule end
-GeneralizedGenerated.NGG.@implement GeneralizedGenerated.NGG.Typeable{PySymPyModule}
-const sympy = PySymPyModule()
-Base.getproperty(::PySymPyModule, s::Symbol) = Base.getproperty(_pysympy, s)
-
-const symfuncs = Dict()
-
-_pow(a,b) = Base.:^(float(a),b)
-
-function __init__()
-    stats = PyCall.pyimport_conda("sympy.stats", "sympy")
-    SymPy.import_from(stats)
-    global stats = stats
-    global _pysympy = SymPy.sympy
-
-
-
-    # for dist in [:Normal, :Cauchy, :Laplace, :Beta, :Uniform]
-    #     @eval begin
-    #         function Distributions.$dist(μ::Sym, σ::Sym)
-    #             println("Evaluating ",$dist, "(",μ," :: Sym, ", σ, " :: Sym)")
-    #             stats.$dist(:dist, μ,σ) |> SymPy.density
-    #         end
-
-    #         Distributions.$dist(μ,σ) = $dist(promote(μ,σ)...)
-    #     end
-    # end
-
-
-
-    # https://discourse.julialang.org/t/pyobjects-as-keys/26521/2
-    merge!(symfuncs, Dict(
-        sympy.log => Base.log
-      , sympy.Pow => _pow
-      , sympy.Abs => Base.abs
-      , sympy.Indexed => getindex
-    ))
-
-    @eval begin
-    @gg M function codegen(_::Type{M}, _m::Model, _args, _data) where M <: TypeLevel{Module}
-        f = _codegen(type2model(_m))
-        Expr(:let,
-            Expr(:(=), :M, from_type(M)),
-            :($f(_args, _data)))
-    end
-end
-
-end
-
-
-struct SymDist
-    logpdf :: Function
-end
-
-logpdf(d::SymDist, x) = d.logpdf(sym(x))
-
-
-# julia> logpdf(SymPy.density(Soss.stats.Poisson(:Poisson,sym(:λ))), sym(:x))
-# x⋅log(λ) - λ - log(x!)
-
-SpecialFunctions.logfactorial(x::Sym) = sympy.loggamma(x+1)
-
-Distributions.Poisson(λ::Sym) = SymDist(x -> x * log(λ) - λ - logfactorial(x))
-Distributions.Bernoulli(p::Sym) = SymDist(y -> y * log(p) + (1-y) * log(1-p))
-
-for dist in [:Bernoulli, :Poisson]
-    @eval begin
-        logpdf(d::$dist, x::Sym) = logpdf($dist(sym.(Distributions.params(d))...), x)
-    end
-end
-
-
-
-# "Half" distributions
-for dist in [:Normal, :Cauchy]
-    let half = Symbol(:Half, dist)
-        @eval begin
-            logpdf(d::$half, x::Sym) = 2 * logpdf($dist(0, sym.(d.σ)), x)
-        end
-    end
-end
-
-for dist in [:Normal, :Cauchy, :Laplace, :Beta, :Uniform]
-    @eval begin
-        function Distributions.$dist(μ::Sym, σ::Sym)
[…]
