+++ Project.toml
@@ -20,7 +20,6 @@ MacroTools = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
 MonteCarloMeasurements = "0987c9cc-fe09-11e8-30f0-b96dd679fdca"
 NamedTupleTools = "d9ec5142-1e00-5aa0-9d6a-321866360f50"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
-PyCall = "438e738f-606a-5dbb-bf0a-cddfbfd45ab0"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 RecipesBase = "3cdcf5f2-1ef4-517c-9805-6587b60abb01"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
@@ -50,7 +49,6 @@ MLStyle = "0.3"
 MacroTools = "0.5"
 MonteCarloMeasurements = "0.6,0.7, 0.8"
 NamedTupleTools = "0.12, 0.13"
-PyCall = "1.91"
 RecipesBase = "0.7,0.8"
 Reexport = "0.2"
 ResumableFunctions = "0.5"
