### 2020-04-06 a9290f8
CompatHelper: bump compat for "RecipesBase" to "1.0"
