+++ Project.toml
@@ -52,7 +52,7 @@ MacroTools = "0.5"
 MonteCarloMeasurements = "0.6,0.7, 0.8"
 NamedTupleTools = "0.12, 0.13"
 PyCall = "1.91"
-RecipesBase = "0.7,0.8"
+RecipesBase = "0.7,0.8, 1.0"
 Reexport = "0.2"
 ResumableFunctions = "0.5"
 ReverseDiff = "1.0"
