### 2020-04-13 #2062 @malmaud
I had paused since it sounds like we're waiting for @cscherrer to add some tests and docs.
