+++ src/Soss.jl
@@ -41,6 +41,7 @@ include("primitives/likelihood-weighting.jl")
 include("transforms/predictive.jl")
 include("transforms/markovblanket.jl")
 include("transforms/do.jl")
+include("transforms/prune.jl")
 
 include("symbolic/symbolic.jl")
 include("symbolic/codegen.jl")
