### 2020-04-28 4d393af
Reconciled arguments
Same pattern as `predictive()` works here as far as I can tell.
