### 2020-04-29 4750541
remove args not depended upon by body
use `simplify = false` for "just prune below" behavior
