+++ src/transforms/prune.jl
@@ -1,7 +1,7 @@
 
 
 export prune
-function prune(m::Model, xs :: Symbol...)
+function prune(m::Model, xs :: Symbol...; simplify = true)
     po = poset(m) #Creates a new SimplePoset, so no need to copy before mutating
 
     newvars = variables(m)
@@ -11,9 +11,15 @@ function prune(m::Model, xs :: Symbol...)
         setdiff!(newvars, [x])
     end
 
-    # Removes unused arguments, removes arguments from variable list.
+    # Keep arguments in newvars
     newargs = arguments(m) ∩ newvars
     setdiff!(newvars, newargs)
+
+    if simplify
+        # keep arguments only if depended upon by newvars
+        dependencies = mapfoldl(var -> below(po, var), vcat, newvars, init = Symbol[]) # mapfoldl needed since newvars can be empty
+        newargs = dependencies ∩ newargs
+    end
 
     theModule = getmodule(m)
     m_init = Model(theModule, newargs, NamedTuple(), NamedTuple(), nothing)
