+++ src/Soss.jl
@@ -42,6 +42,8 @@ include("transforms/predictive.jl")
 include("transforms/markovblanket.jl")
 include("transforms/do.jl")
 include("transforms/prune.jl")
+include("transforms/prior.jl")
+include("transforms/prune.jl")
 
 include("symbolic/symbolic.jl")
 include("symbolic/codegen.jl")
