### 2020-05-02 968aa67
first doctests
