### 2020-05-02 cd1a6f9
first doctests
