+++ test/runtests.jl
@@ -1,13 +1,20 @@
 using Soss
 using Test
 using Weave
+using Documenter
 
-function buildREADME() 
+DocMeta.setdocmeta!(Soss, :DocTestSetup, :(using Soss); recursive=true)
+
+function buildREADME()
     weave("../README.jmd", doctype= "github", throw_errors=true, cache=:refresh, args=Dict(:seed => 6))
     return true
 end
 
 # write your own tests here
-@testset "README" begin
-    @test buildREADME()
+#@testset "README" begin
+#    @test buildREADME()
+#end
+
+@testset "Doctests" begin
+    doctest(Soss)
 end
