+++ README.md
@@ -185,23 +185,14 @@ end;
 
 julia> 
 symlogpdf(m2)
-                                                    N                         
-                                                   ___                        
-                                                   ╲                          
-                                                    ╲                         
--0.918938533204673⋅N - 0.918938533204673⋅k - 0.5⋅   ╱    (y[_j1] - 1.0⋅yhat[_j
-                                                   ╱                          
-                                                   ‾‾‾                        
-                                                 _j1 = 1                      
-
-              k           
-             ___          
-             ╲            
-   2          ╲          2
-1])  - 0.5⋅   ╱    β[_j1] 
-             ╱            
-             ‾‾‾          
-           _j1 = 1        
+                                                    N                                       k           
+                                                   ___                                     ___          
+                                                   ╲                                       ╲            
+                                                    ╲                            2          ╲          2
+-0.918938533204673⋅N - 0.918938533204673⋅k - 0.5⋅   ╱    (y[_j1] - 1.0⋅yhat[_j1])  - 0.5⋅   ╱    β[_j1] 
+                                                   ╱                                       ╱            
+                                                   ‾‾‾                                     ‾‾‾          
+                                                 _j1 = 1                                 _j1 = 1        
 
 ````
 
@@ -213,23 +204,14 @@ There's clearly some redundant computation within the sums, so it helps to expan
 
 ````julia
 julia> symlogpdf(m2) |> expandSums |> foldConstants
-                                                    N                         
-                                                   ___                        
-                                                   ╲                          
-                                                    ╲                         
--0.918938533204673⋅N - 0.918938533204673⋅k - 0.5⋅   ╱    (y[_j1] - 1.0⋅yhat[_j
-                                                   ╱                          
-                                                   ‾‾‾                        
-                                                 _j1 = 1                      
-
-              k           
-             ___          
-             ╲            
-   2          ╲          2
-1])  - 0.5⋅   ╱    β[_j1] 
-             ╱            
-             ‾‾‾          
-           _j1 = 1        
+                                                    N                                       k           
+                                                   ___                                     ___          
+                                                   ╲                                       ╲            
+                                                    ╲                            2          ╲          2
+-0.918938533204673⋅N - 0.918938533204673⋅k - 0.5⋅   ╱    (y[_j1] - 1.0⋅yhat[_j1])  - 0.5⋅   ╱    β[_j1] 
+                                                   ╱                                       ╱            
+                                                   ‾‾‾                                     ‾‾‾          
+                                                 _j1 = 1                                 _j1 = 1        
 
 ````
 
@@ -248,7 +230,7 @@ julia>
 -15.84854642585797
 
 julia> @btime logpdf($m2(X=X), $truth, $codegen)
-  288.658 ns (5 allocations: 208 bytes)
+  306.753 ns (5 allocations: 208 bytes)
 -15.848546425857968
 
 ````
