### 2020-05-02 559ca3a
first doctests
