+++ src/core/jointdistribution.jl
@@ -1,5 +1,9 @@
+using Distributions: ValueSupport, VariateForm
 
-struct JointDistribution{A0,A,B,M}
+struct MixedSupport <: ValueSupport end
+struct MixedVariate <: VariateForm end
+
+struct JointDistribution{A0,A,B,M} <: Distribution{MixedVariate, MixedSupport}
     model::Model{A,B,M}
     args::A0
 end
