### 2020-05-04 fe6d713
Remove Documenter from main Project.toml
