### 2020-05-05 05346e6
get `using Soss` into doctests (thanks vasa)
