### 2020-05-05 57a7d7e
Have Travis build `dev`
