### 2020-05-05 a180518
Update Julia for Travis config
