+++ .travis.yml
@@ -3,7 +3,7 @@ os:
     - linux
     - osx
 julia:
-    - 1.3
+    - 1.4
     - nightly
     
 branches:
