### 2020-05-07 ee532dc
Changed argument from simplify to trim_args
