+++ src/distributions/markovchain.jl
@@ -7,6 +7,7 @@ export MarkovChain
 
 `MarkovChain(pars, step)` defines a Markov Chain with global parameters `pars` and transition kernel `step`. Here, `pars` is a named tuple, and `step` is a Soss model that takes arguments `(pars, state)` and returns a `next` value containing the new `pars` and `state`.
 
+NOTE: This is experimental, and may change in the near future.
 
 mstep = @model pars,state begin
     σ = pars.σ
