### 2020-05-08 e37ac0a
CompatHelper: add new compat entry for "Setfield" at version "0.6"
