+++ Project.toml
@@ -57,6 +57,7 @@ RecipesBase = "0.7,0.8"
 Reexport = "0.2"
 ResumableFunctions = "0.5"
 ReverseDiff = "1.0"
+Setfield = "0.6"
 SimpleGraphs = "0.3"
 SimplePartitions = "0.2"
 SimplePosets = "0.0"
