### 2020-05-09 bdbc3ba
CompatHelper: bump compat for "MLStyle" to "0.4"
