+++ Project.toml
@@ -48,7 +48,7 @@ Graphs = "0.10"
 IterTools = "1.3"
 LazyArrays = "0.14, 0.15, 0.16"
 LogDensityProblems = "0.10"
-MLStyle = "0.3"
+MLStyle = "0.3, 0.4"
 MacroTools = "0.5"
 MonteCarloMeasurements = "0.6,0.7, 0.8"
 NamedTupleTools = "0.12, 0.13"
