+++ src/core/model.jl
@@ -1,192 +1,0 @@
-export Model, @model
-using MLStyle
-
-using MacroTools: @q, striplines
-using SimpleGraphs
-using SimplePosets
-using GeneralizedGenerated
-
-struct Model{A,B,M} 
-    args  :: Vector{Symbol}
-    vals  :: NamedTuple
-    dists :: NamedTuple
-    retn  :: Union{Nothing, Symbol, Expr}
-end
-
-argstype(::Model{A,B,M}) where {A,B,M} = A
-bodytype(::Model{A,B,M}) where {A,B,M} = B
-
-argstype(::Type{Model{A,B,M}}) where {A,B,M} = A
-bodytype(::Type{Model{A,B,M}}) where {A,B,M} = B
-
-getmodule(::Type{Model{A,B,M}}) where {A,B,M} = from_type(M)
-getmodule(::Model{A,B,M}) where {A,B,M} = from_type(M)
-
-getmoduletypencoding(::Type{Model{A,B,M}}) where {A, B, M} = M
-getmoduletypencoding(::Model{A,B,M}) where {A,B,M} = M
-
-function Model(theModule::Module, args, vals, dists, retn)
-    M = to_type(theModule)
-    A = NamedTuple{Tuple(args)}
-    m = Model{A,Any,M}(args, vals, dists, retn)
-    B = convert(Expr, m).args[end] |> to_type
-    Model{A,B,M}(args, vals, dists, retn)
-end
-
-function type2model(::Type{Model{A,B,M}}) where {A,B,M}
-    args = [fieldnames(A)...]
-    body = from_type(B)
-    Model(from_type(M), convert(Vector{Symbol},args), body)
-end
-
-function emptyModel(theModule::Module)
-    M = to_type(theModule)
-    A = NamedTuple{(),Tuple{}}                    
-    B = (@q begin end) |> to_type
-    Model{A,B,M}([], NamedTuple(), NamedTuple(), nothing)
-end
-
-
-function Base.merge(m1::Model, m2::Model) 
-    theModule = getmodule(m1)
-    @assert theModule == getmodule(m2)
-    vals = merge(m1.vals, m2.vals)
-    args = setdiff(union(m1.args, m2.args), keys(vals))
-    dists = merge(m1.dists, m2.dists)
-    retn = maybesomething(m2.retn, m1.retn) # m2 first so it gets priority
-  
-    Model(theModule, args, vals, dists, retn)
-end
-
-Base.merge(m::Model, ::Nothing) = m
-
-Model(theModule::Module, st::Assign) = Model(theModule, Symbol[], namedtuple(st.x)([st.rhs]), NamedTuple(), nothing)
-Model(theModule::Module, st::Sample) = Model(theModule, Symbol[], NamedTuple(), namedtuple(st.x)([st.rhs]), nothing)
-Model(theModule::Module, st::Return) = Model(theModule, Symbol[], NamedTuple(), NamedTuple(), st.rhs)
-Model(theModule::Module, st::LineNumber) = emptyModel(theModule)
-
-Model(theModule::Module, ::LineNumberNode) = emptyModel(theModule)
-
-function Model(theModule::Module, expr :: Expr)
-    nt = NamedTuple()
-    @match expr begin
-        :($k = $v)   => Model(theModule, Assign(k,v))
-        :($k ~ $v)   => Model(theModule, Sample(k,v))
-        Expr(:return, x...) => Model(theModule, Return(x[1]))
-        Expr(:block, body...) => foldl(merge, map(body) do line Model(theModule, line) end)
-        :(@model $lnn $body) => Model(theModule, body)
-        :(@model $lnn $args $body) => Model(theModule, args.args, body)
-
[…]
