### 2020-05-09 71e6852
MarkovChain example (#140)
* version++

* Add Setfield

* Markov chains

* tidying up

* include markovchain.jl

* prune

* Reconciled arguments

Same pattern as `predictive()` works here as far as I can tell.

* remove args not depended upon by body

use `simplify = false` for "just prune below" behavior

* prior

* include `prior` and `prune`

* remove duplicate include (oops)

* update travis

* first doctests

* accidentally committed this..

Forgot to undo this.

* Remove Documenter from main Project.toml

* get `using Soss` into doctests (thanks vasa)

* Have Travis build `dev`

* Update
[…]
