+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"misc/#Models-and-JointDistributions-1","page":"Models and JointDistributions","title":"Models and JointDistributions","text":"","category":"section"},{"location":"misc/#","page":"Models and JointDistributions","title":"Models and JointDistributions","text":"A Model in Soss","category":"page"},{"location":"misc/#Model-Combinators-1","page":"Models and JointDistributions","title":"Model Combinators","text":"","category":"section"},{"location":"misc/#Building-Inference-Algorithms-1","page":"Models and JointDistributions","title":"Building Inference Algorithms","text":"","category":"section"},{"location":"misc/#Inference-Primitives-1","page":"Models and JointDistributions","title":"Inference Primitives","text":"","category":"section"},{"location":"misc/#","page":"Models and JointDistributions","title":"Models and JointDistributions","text":"At its core, Soss is about source code generation. Instances of this are referred to as inference primitives, or simply \"primitives\". As a general rule, new primitives are rarely needed. A wide variety of inference algorithms can be built using what's provided. ","category":"page"},{"location":"misc/#","page":"Models and JointDistributions","title":"Models and JointDistributions","text":"To easily find all available inference primitives, enter Soss.source<TAB> at a REPL. Currently this returns this result:","category":"page"},{"location":"misc/#","page":"Models and JointDistributions","title":"Models and JointDistributions","text":"julia> Soss.source\nsourceLogpdf         sourceRand            sourceXform\nsourceParticles      sourceWeightedSample","category":"page"},{"location":"misc/#","page":"Models and JointDistributions","title":"Models and JointDistributions","text":"The general pattern is that a primitive sourceFoo specifies how code is generated for an inference function foo. ","category":"page"},{"location":"misc/#","page":"Models and JointDistributions","title":"Models and JointDistributions","text":"For more details on inference primitives, see the Internals section.","category":"page"},{"location":"misc/#Inference-Functions-1","page":"Models and JointDistributions","title":"Inference Functions","text":"","category":"section"},{"location":"misc/#","page":"Models and JointDistributions","title":"Models and JointDistributions","text":"An inference function is a function that takes a JointDistribution as an argument, and calls at least one inference primitive (not necessarily directly). The wrapper around each primitive is a special case of this, but most inference functions work at a higher level of abstraction.","category":"page"},{"location":"misc/#","page":"Models and JointDistributions","title":"Models and JointDistributions","text":"There's some variability , but is often of the form","category":"page"},{"location":"misc/#","page":"Models and JointDistributions","title":"Models and JointDistributions","text":"foo(d::JointDistribution, data::NamedTuple)","category":"page"},{"location":"misc/#","page":"Models and JointDistributions","title":"Models and JointDistributions","text":"For example, advancedHMC uses TuringLang/AdvancedHMC.jl , which needs a logpdf and its gradient. ","category":"page"},{"location":"misc/#","page":"Models and JointDistributions","title":"Models and JointDistributions","text":"Most inference algorithms can be expressed in terms of inference primitives. ","category":"page"},{"location":"misc/#Chain-Combinators-1","pa
[…]
