+++ src/distributions/markovchain.jl
@@ -31,7 +31,9 @@ end
 
 # output
 
-
+(x = -6.596883394256064,)
+(x = 0.48200039561318864,)
+(x = -2.838556784903994,)
 ```
 """
 struct MarkovChain{P,D}
