### 2020-05-10 13d800c
logpdf -> logdensity
