+++ src/symbolic/symbolic.jl
@@ -64,13 +64,13 @@ end
 
 
 struct SymDist
-    logpdf :: Function
+    logdensity :: Function
 end
 
-logpdf(d::SymDist, x) = d.logpdf(sym(x))
+logdensity(d::SymDist, x) = d.logdensity(sym(x))
 
 
-# julia> logpdf(SymPy.density(Soss.stats.Poisson(:Poisson,sym(:λ))), sym(:x))
+# julia> logdensity(SymPy.density(Soss.stats.Poisson(:Poisson,sym(:λ))), sym(:x))
 # x⋅log(λ) - λ - log(x!)
 
 SpecialFunctions.logfactorial(x::Sym) = sympy.loggamma(x+1)
@@ -80,7 +80,7 @@ Distributions.Bernoulli(p::Sym) = SymDist(y -> y * log(p) + (1-y) * log(1-p))
 
 for dist in [:Bernoulli, :Poisson]
     @eval begin
-        logpdf(d::$dist, x::Sym) = logpdf($dist(sym.(Distributions.params(d))...), x)
+        logdensity(d::$dist, x::Sym) = logdensity($dist(sym.(Distributions.params(d))...), x)
     end
 end
 
@@ -90,7 +90,7 @@ end
 for dist in [:Normal, :Cauchy]
     let half = Symbol(:Half, dist)
         @eval begin
-            logpdf(d::$half, x::Sym) = 2 * logpdf($dist(0, sym.(d.σ)), x)
+            logdensity(d::$half, x::Sym) = 2 * logdensity($dist(0, sym.(d.σ)), x)
         end
     end
 end
@@ -230,9 +230,9 @@ end
 export score
 score(ℓ::Sym, v) = diff(ℓ, sym(v))
 
-score(m::Model, v) = score(symlogpdf(m), v)
+score(m::Model, v) = score(symlogdensity(m), v)
 
-marginal(m::Model, v) = marginal(m |> symlogpdf, v)
+marginal(m::Model, v) = marginal(m |> symlogdensity, v)
 
 symvar(st) = :($sympy.IndexedBase($(st.x)))
 
@@ -242,8 +242,8 @@ function symvar(st::Sample)
 end
     
 
-export sourceSymlogpdf
-function sourceSymlogpdf()
+export sourceSymlogdensity
+function sourceSymlogdensity()
     function(_m::Model)
         function proc(_m, st :: Assign)
             x = st.x
@@ -252,7 +252,7 @@ function sourceSymlogpdf()
         end
 
         function proc(_m, st :: Sample)
-            s = :(_ℓ += symlogpdf($(st.rhs), $(symvar(st))))
+            s = :(_ℓ += symlogdensity($(st.rhs), $(symvar(st))))
             end
         proc(_m, st :: Return)     = nothing
         proc(_m, st :: LineNumber) = nothing
@@ -301,13 +301,13 @@ function For(f::F, θ::NTuple{N,Sym}) where {F, N}
 end
 
 
-# logpdf(d::For{F,N,T}, x::Array{Symbol, N}) where {F,N,T} = logpdf(d,sym.(x))
+# logdensity(d::For{F,N,T}, x::Array{Symbol, N}) where {F,N,T} = logdensity(d,sym.(x))
 
-export symlogpdf
-function symlogpdf(d::For{F,T,D,X}, x::Sym) where {F, N, J <: Union{Sym,Integer}, T <: NTuple{N,J}, D,  X}
+export symlogdensity
[…]
