### 2020-05-11 c1d8c12
Initial doctest for `advancedHMC`
Simple docstring and test for the `advancedHMC` function. Notes:
  - Raw prefix to docstring is necessary to correctly parse the output filter.
  - E[β|x] is not deterministic, so it currently only checks the first two significant digits.
