### 2020-05-11 bb087d3
finely control output generated by doctest
Switched from attempting to filter generated out, to switching off sampling progress printting and deprecation warnings.
