+++ .travis.yml
@@ -32,5 +32,5 @@ jobs:
       script:
         - julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd()));
                                                Pkg.instantiate()'
-        - julia --project=docs/ docs/make.jl
+        - julia --depwarn=no --project=docs/ docs/make.jl
       after_success: skip
