### 2020-05-11 e627439
Change to flagging deprecation warnings
Revert the change to the travis config, and update the `advancedHMC` sampling call based on their current docs.
