### 2020-05-11 df3d2b1
logpdf -> logdensity
