+++ Project.toml
@@ -6,7 +6,6 @@ version = "0.11.0"
 [deps]
 AdvancedHMC = "0bf59076-c3b1-5ca4-86bd-e02cd72cde3d"
 DiffResults = "163ba53b-c6d8-5494-b064-1a9d43ac40c5"
-Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 DynamicHMC = "bbc10e6e-7c05-544b-b16e-64fede858acb"
 FillArrays = "1a297f60-69ca-5386-bcde-b61e274b549b"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
@@ -17,8 +16,10 @@ LazyArrays = "5078a376-72f3-5289-bfd5-ec5146d43c02"
 LogDensityProblems = "6fdf6af0-433a-55f7-b3ed-c6c6e0b8df7c"
 MLStyle = "d8e11817-5142-5d16-987a-aa16d5891078"
 MacroTools = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
+Measures = "eadaa1a4-d27c-401d-8699-e962e1bbc33b"
 MonteCarloMeasurements = "0987c9cc-fe09-11e8-30f0-b96dd679fdca"
 NamedTupleTools = "d9ec5142-1e00-5aa0-9d6a-321866360f50"
+Permutations = "2ae35dd2-176d-5d53-8349-f30d82d94d4f"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 PyCall = "438e738f-606a-5dbb-bf0a-cddfbfd45ab0"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
@@ -40,7 +41,6 @@ TransformVariables = "84d833dd-6860-57f9-a1a7-6da5db126cff"
 [compat]
 AdvancedHMC = "0.2"
 DiffResults = "1.0"
-Distributions = "0.21, 0.22, 0.23"
 DynamicHMC = "2.1"
 FillArrays = "0.8"
 ForwardDiff = "0.10"
