### 2020-05-11 aadab90
CompatHelper: bump compat for "RecipesBase" to "1.0" (#129)
* CompatHelper: bump compat for "RecipesBase" to "1.0"

* README

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Chad Scherrer <chad.scherrer@gmail.com>
