### 2020-05-12 d85c942
initial docstring and doctest for `dynamicHMC`
This follows the docstring and doctest for `advancedHMC`. The `method` argument is undocumented because I couldn't figure out exactly what it is doing.
