+++ src/inference/dynamicHMC.jl
@@ -9,6 +9,69 @@ import LogDensityProblems: ADgradient
 
 export dynamicHMC
 
+"""
+    dynamicHMC(
+        rng::AbstractRNG,
+        m::JointDistribution,
+        _data,
+        N::Int = 1000;
+        method = logpdf,
+        ad_backend = Val(:ForwardDiff),
+        reporter = DynamicHMC.NoProgressReport(),
+        kwargs...)
+    
+
+Draw `N` samples from the posterior distribution of parameters defined in Soss model `m`, conditional on `_data`. Samples are drawn using Hamiltonial Monte Carlo (HMC) from the `DynamicHMC.jl` package.
+
+This function is essentially a wrapper around `DynamicHMC.mcmc_with_warmup()`. Arguments `reporter`, `ad_backend` [DynamicHMC docs here](https://tamaspapp.eu/DynamicHMC.jl/stable/interface/#DynamicHMC.mcmc_with_warmup))
+
+## Arguments
+* `rng`: Random number generator.
+* `m`: Soss model.
+* `_data`: `NamedTuple` of data to condition on.
+
+
+## Keyword Arguments
+*  `N = 1000`: Number of samples to draw.
+*  `method = logpdf`: .
+*  `ad_backend = Val(:ForwardDiff)`: Automatic differentiation backend.
+*  `reporter = DynamicHMC.NoProgressReport()`: Specify logging during sampling. Default: do not log progress.
+*  `kwargs`: Additional keyword arguments passed to core sampling function `DynamicHMC.mcmc_with_warmup()`.
+
+
+Returns an Array of `Namedtuple` of length `N`. Each entry in the array is a sample of parameters indexed by the parameter symbol.
+
+## Example
+
+```jldoctest
+
+using Random
+Random.seed!(42);
+rng = MersenneTwister(42);
+
+m = @model x begin
+    β ~ Normal()
+    yhat = β .* x
+    y ~ For(eachindex(x)) do j
+        Normal(yhat[j], 2.0)
+    end
+end
+
+x = randn(50);
+truth = rand(m(x=x));
+
+post = dynamicHMC(rng, m(x=x), (y=truth.y,));
+E_β = mean(getfield.(post, :β))
+
+println("true β: " * string(round(truth.β, digits=2)))
+println("Posterior mean β: " * string(round(E_β, digits=2)))
+
+# output
+true β: 0.3
+Posterior mean β: 0.47
+```
+
+"""
 function dynamicHMC(
     rng::AbstractRNG,
     m::JointDistribution,
