### 2020-05-12 ecf9213
Update src/inference/dynamicHMC.jl
Co-authored-by: Chad Scherrer <chad.scherrer@gmail.com>
