### 2020-05-12 8621448
Initial docstrings and doctests for inference methods (#142)
* Initial doctest for `advancedHMC`

Simple docstring and test for the `advancedHMC` function. Notes:
  - Raw prefix to docstring is necessary to correctly parse the output filter.
  - E[β|x] is not deterministic, so it currently only checks the first two significant digits.

* finely control output generated by doctest

Switched from attempting to filter generated out, to switching off sampling progress printting and deprecation warnings.

* Change to flagging deprecation warnings

Revert the change to the travis config, and update 
[…]
