### 2020-05-13 8d29ba8
importing specific functions from SimpleGraphs
