### 2020-05-13 b6bc646
Update src/core/toposort.jl
using instead of import

Co-authored-by: Chad Scherrer <chad.scherrer@gmail.com>
