### 2020-05-13 ce97f25
importing specific functions from SimpleGraphs to fix name clashes (#143)
* using specific functions from SimpleGraphs
