### 2020-05-15 d45d2df
Copying Distribution's conventions for one-parameter specifications of these distributions
