### 2020-05-17 38bf2da
Fix symbolics in README
