### 2020-05-17 66ffb46
Added transformation tests, refactored runtests to `include` testsets
