### 2020-05-17 b3f90a2
Added transformation tests, refactored runtests to `include` testsets
