### 2020-05-17 4a025b3
Simplify, refactor transformations to share more code.
