### 2020-05-17 ca20c69
Simplify, refactor transformations to share more code.
