### 2020-05-17 783c622
uncomment readme.jl test
