+++ test/runtests.jl
@@ -2,5 +2,5 @@ using Soss
 using Test
 
 include("doctests.jl")
-#include("readme.jl")
+include("readme.jl")
 include("transforms.jl")
