+++ src/Soss.jl
@@ -38,7 +38,7 @@ include("primitives/likelihood-weighting.jl")
 # include("primitives/bijectors.jl")
 
 
-include("transforms/predictive.jl")
+include("transforms/predict.jl")
 include("transforms/markovblanket.jl")
 include("transforms/basictransforms.jl")
 
