### 2020-05-17 f54949c
redundant setdiff
