### 2020-05-17 9df005e
Correct evalf stuff in SymPy
