### 2020-05-18 0b3231f
trim_args took a bad shortcut. Should work as desired now
