### 2020-05-18 b29845c
trim_args took a bad shortcut. Should work as desired now
