### 2020-05-20 #2062 @arfon
:wave: @cscherrer - have you managed to make any progress since April? JOSS is now reopen for submissions so we're actively trying to wrap up existing reviews.

:wave: @malmaud - do you think you might be able to start work on your review sometime in the next few weeks?
