### 2020-05-20 #2062 @cscherrer
Thanks @arfon and @malmaud . We've been digging into the Julia documentation system, and now have a good start on documentation (previously it had been just the README). I'm hoping to have this ready to move forward within a few weeks. We can keep this paused until then. Please let me know if we need a hard deadline or this delay causes some problem for you; I could instead cancel and re-apply when we're ready if that is somehow better for you.
