### 2020-05-20 1cf635a
finished before/after. Added docs for API transforms.
