+++ src/Soss.jl
@@ -40,8 +40,8 @@ include("primitives/likelihood-weighting.jl")
 
 include("transforms/predict.jl")
 include("transforms/markovblanket.jl")
+include("transforms/utils.jl")
 include("transforms/basictransforms.jl")
-include("transforms/basictransforms2.jl")
 
 include("symbolic/symbolic.jl")
 include("symbolic/codegen.jl")
