### 2020-05-20 fd64093
finished before/after. Added docs for API transforms.
