+++ src/transforms/utils.jl
@@ -1,0 +1,133 @@
+using SimplePartitions: find_part
+import SimpleGraphs
+using SimpleGraphs: vlist, out_neighbors
+
+# Type piracy, this should really go in SimpleGraphs
+# OTOH there's not anything else this could really mean
+# function SimpleGraphs.components(g::SimpleDigraph)
+#     SimpleGraphs.components(convert(SimpleGraph, g))
+# end
+
+function _before(g::SimpleDigraph, v)
+    parents = in_neighbors(g, v)
+    for i in parents
+        append!(parents, _before(g, i))
+    end
+    return parents
+end
+
+before(g::SimpleDigraph, v; inclusive = true) = inclusive ? push!(_before(g, v), v) : _before(g, v)
+function before(g::SimpleDigraph, vs...; inclusive = true)
+    parents = inclusive ? collect(vs) : Symbol[]
+    for v in vs
+        append!(parents, before(g, v, inclusive = inclusive))
+    end
+    if !inclusive
+        setdiff!(parents, vs)
+    end
+    return unique!(parents)
+end
+
+notbefore(g::SimpleDigraph, vs...; inclusive = false) = setdiff(vlist(g), before(g, vs...; inclusive = !inclusive))
+
+function _after(g::SimpleDigraph, v)
+    children = out_neighbors(g, v)
+    for i in children
+        append!(children, _after(g, i))
+    end
+    return children
+end
+
+after(g::SimpleDigraph, v; inclusive = true) = inclusive ? push!(_after(g, v), v) : _after(g, v)
+
+function after(g::SimpleDigraph, vs...; inclusive = true)
+    children = inclusive ? collect(vs) : Symbol[]
+    for v in vs
+        append!(children, after(g, v, inclusive = inclusive))
+    end
+    if !inclusive
+        setdiff!(children, vs)
+    end
+    return unique!(children)
+end
+
+notafter(g::SimpleDigraph, vs...; inclusive = false) = setdiff(vlist(g), after(g, vs...; inclusive = !inclusive))
+
+function assemblefrom(m::Model, params, args)
+    theModule = getmodule(m)
+    m_init = Model(theModule, args, NamedTuple(), NamedTuple(), nothing)
+    m = foldl(params; init=m_init) do m0,v
+        merge(m0, Model(theModule, findStatement(m, v)))
+    end
+    return m
+end
+
+"""
+Enforces that arguments are parents of params (and not params themselves).
+"""
+function trim_args!(args, m, params)
+    g = digraph(m)
+    intersect!(args, setdiff(vcat((parents(g, p) for p in params)...), params))
+end
+
+sourcenodes(g::SimpleDigraph) = setdiff(vlist(g), [last(e) for e in elist(g)])
+sinknodes(g::SimpleDigraph) = setdiff(vlist(g), [first(e) for e in elist(g)])
+
+function after(m::Model, xs...; strict = false)
+    g = digraph(m)
+    for x in xs
+        map(a->delete!(g, a, x), in_neighbors(g, x))
+    end
+    parms = Symbol[]
+    for v in sinknodes(g)
+        sinkparents = before(g, v)
+        for x in xs
+            if !strict || x in sinkparents
+                append!(parms, sinkparents)
+            end
+        end
+    end
+    args = arguments(m) ∪ xs # Will trim later.
+    parms = setdiff(parms, args)
+    trim_args!(args, m, parms)
+    return assemblefrom(m, parms, args)
+end
+
+export before
+"""
+    before(m::Model, xs...; inclusive=true, strict=true)
+
[…]
