### 2020-05-22 631d631
update readme (NOTE PERFORMANCE REGRESSIONS)
