+++ README.md
@@ -47,6 +47,16 @@ end;
 ````
 
 
+````
+@model X begin
+        β ~ Normal() |> iid(size(X, 2))
+        y ~ For(eachrow(X)) do x
+                Normal(x' * β, 1)
+            end
+    end
+````
+
+
 
 
 
@@ -62,6 +72,7 @@ Just a few of the things we can do in Soss:
 Let's use our model to build some fake data:
 ````julia
 julia> import Random; Random.seed!(3)
+Random.MersenneTwister(UInt32[0x00000003], Random.DSFMT.DSFMT_state(Int32[-1359582567, 1073454075, 1934390716, 1073583786, -114685834, 1073112842, -1913218479, 1073122729, -73577195, 1073266439  …  1226759590, 1072980451, -1366384707, 1073012992, 1661148031, 2121090155, 141576524, -658637225, 382, 0]), [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0  …  0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0], UInt128[0x00000000000000000000000000000000, 0x00000000000000000000000000000000, 0x00000000000000000000000000000000, 0x00000000000000000000000000000000, 0x00000000000000000000000000000000, 0x00000000000000000000000000000000, 0x00000000000000000000000000000000, 0x00000000000000000000000000000000, 0x00000000000000000000000000000000, 0x00000000000000000000000000000000  …  0x00000000000000000000000000000000, 0x00000000000000000000000000000000, 0x00000000000000000000000000000000, 0x00000000000000000000000000000000, 0x00000000000000000000000000000000, 0x00000000000000000000000000000000, 0x00000000000000000000000000000000, 0x00000000000000000000000000000000, 0x00000000000000000000000000000000, 0x00000000000000000000000000000000], 1002, 0)
 
 julia> X = randn(6,2)
 6×2 Array{Float64,2}:
@@ -78,6 +89,7 @@ julia> X = randn(6,2)
 
 ````julia
 julia> truth = rand(m(X=X));
+(X = [1.1915557734285787 0.10079289135480324; -2.5197330871745263 -0.0019741367391015213; … ; -0.1016067940589428 1.158074626662026; -1.5425131978228126 -0.47515878362112707], β = [0.07187269298745927, -0.5128103336795292], y = [0.10079289135480324, -2.5197330871745263, 2.0748097755419757, 0.8442227439533416, 1.158074626662026, -0.47515878362112707])
 
 julia> pairs(truth)
 pairs(::NamedTuple) with 3 entries:
@@ -120,6 +132,27 @@ Often these are easier to work with in terms of `particles` (built using [MonteC
 
 ````julia
 julia> post = dynamicHMC(m(X=truth.X), (y=truth.y,));
+1000-element Array{NamedTuple{(:β,),Tuple{Array{Float64,1}}},1}:
+ (β = [0.39734466526796725, 0.9980730158207407],)
+ (β = [0.45764947899215946, 0.7167597584561338],)
+ (β = [0.44310994230948264, 1.1912646425584188],)
+ (β = [0.5804128915563641, 0.9052672328462603],)
+ (β = [0.4911904823400375, 1.3336636471494507],)
+ (β = [0.49901795646365543, 1.3937610114041836],)
+ (β = [0.45240830115259073, 1.0845756727727063],)
+ (β = [0.5850340074895557, 0.9928101856844366],)
+ (β = [0.3914145275648779, 1.0526332372994567],)
[…]
