### 2020-05-24 4f5f5c1
Define (m::Model)()
