### 2020-05-24 a2dff17
Define predictive on empty named tuple
