### 2020-05-24 39baa22
Have `rand` only return sampled values
