### 2020-05-27 bf2a562
Change behavior of calling Models and JoinDistributions
