### 2020-05-27 e14c376
remove redundant method
