### 2020-05-27 504903f
Make `rand` output parameter values
