### 2020-05-27 0422334
helper functions for `source*`
