+++ src/primitives/xform.jl
@@ -11,7 +11,7 @@ using Distributions
 export xform
 
 
-function xform(m::JointDistribution{A, B}, _data = m.args) where {A,B}
+function xform(m::JointDistribution{A, B}, _data = NamedTuple{}()) where {A,B}
     return _xform(getmoduletypencoding(m.model), m.model, m.args, _data)
 end
 
@@ -27,6 +27,8 @@ end
 
 
 export sourceXform
+
+sourceXform(m::Model) = sourceXform()(m)
 
 function sourceXform(_data=NamedTuple())
     function(_m::Model)
