### 2020-05-27 5e6a077
fix doctest to use unexported before, after
