### 2020-05-29 e2390b9
unexporting prior
don't want to take up the user's namespace
