+++ test/transforms.jl
@@ -14,16 +14,15 @@ m = @model (n,α,β) begin
 end
 
 @testset "prior" begin
-    using Soss.prior
-    m1 = prior(m, :x)
-    @test prior(m, :x) ≊ @model (n,α,β) begin
+    m1 = Soss.prior(m, :x)
+    @test Soss.prior(m, :x) ≊ @model (n,α,β) begin
         p ~ Beta(α, β)
         x ~ Binomial(n, p)
     end
-    @test prior(m1, :p) ≊ @model (α, β) begin
+    @test Soss.prior(m1, :p) ≊ @model (α, β) begin
         p ~ Beta(α, β)
     end
-    @test prior(m, :z) ≊ @model (n, α, β) begin
+    @test Soss.prior(m, :z) ≊ @model (n, α, β) begin
         z ~ Binomial(n, α / (α + β))
     end
 end
