### 2020-06-01 0d984ce
CompatHelper: add new compat entry for "Requires" at version "1.0"
