### 2020-06-01 c06a18f
CompatHelper: add new compat entry for "Requires" at version "1.0"
