+++ Project.toml
@@ -56,6 +56,7 @@ NamedTupleTools = "0.12, 0.13"
 PyCall = "1.91"
 RecipesBase = "0.7,0.8, 1.0"
 Reexport = "0.2"
+Requires = "1.0"
 ResumableFunctions = "0.5"
 ReverseDiff = "1.0"
 Setfield = "0.6"
