+++ Project.toml
@@ -51,7 +51,7 @@ LazyArrays = "0.14, 0.15, 0.16"
 LogDensityProblems = "0.10"
 MLStyle = "0.3, 0.4"
 MacroTools = "0.5"
-MonteCarloMeasurements = "0.6,0.7, 0.8"
+MonteCarloMeasurements = "0.6,0.7, 0.8, 0.9"
 NamedTupleTools = "0.12, 0.13"
 PyCall = "1.91"
 RecipesBase = "0.7,0.8, 1.0"
