### 2020-06-28 5414437
CompatHelper: bump compat for "SimplePartitions" to "0.3"
