+++ Project.toml
@@ -60,7 +60,7 @@ ResumableFunctions = "0.5"
 ReverseDiff = "1.0"
 Setfield = "0.6"
 SimpleGraphs = "0.3"
-SimplePartitions = "0.2"
+SimplePartitions = "0.2, 0.3"
 SimplePosets = "0.0"
 SpecialFunctions = "0.10"
 StatsFuns = "0.9"
