### 2020-07-01 c0df704
CompatHelper: bump compat for "SimplePosets" to "0.1"
