+++ Project.toml
@@ -62,7 +62,7 @@ ReverseDiff = "1.0"
 Setfield = "0.6, 0.7"
 SimpleGraphs = "0.3"
 SimplePartitions = "0.2"
-SimplePosets = "0.0"
+SimplePosets = "0.0, 0.1"
 SpecialFunctions = "0.10"
 StatsFuns = "0.9"
 SymPy = "1.0"
