### 2020-07-01 249617b
CompatHelper: bump compat for "SimplePosets" to "0.1"
