+++ Project.toml
@@ -61,7 +61,7 @@ ReverseDiff = "1.0"
 Setfield = "0.6"
 SimpleGraphs = "0.3"
 SimplePartitions = "0.2"
-SimplePosets = "0.0"
+SimplePosets = "0.0, 0.1"
 SpecialFunctions = "0.10"
 StatsFuns = "0.9"
 SymPy = "1.0"
