### 2020-07-01 7697907
CompatHelper: bump compat for "SimpleGraphs" to "0.4"
