+++ Project.toml
@@ -59,7 +59,7 @@ Reexport = "0.2"
 ResumableFunctions = "0.5"
 ReverseDiff = "1.0"
 Setfield = "0.6"
-SimpleGraphs = "0.3"
+SimpleGraphs = "0.3, 0.4"
 SimplePartitions = "0.2"
 SimplePosets = "0.0"
 SpecialFunctions = "0.10"
