### 2020-07-10 2bb54a3
abstract array method for For
