+++ src/distributions/for.jl
@@ -8,34 +8,37 @@ export logpdf
 export rand
 
 export For
-struct For{F,T,D,X} 
-    f :: F  
-    θ :: T
+struct For{F,T,D,X}
+    f::F
+    θ::T
 end
 
 #########################################################
 # T <: NTuple{N,J} where {J <: Integer}
 #########################################################
 
-For(f, θ::J...) where {J <: Integer} = For(f,θ)
+For(f, θ::J...) where {J<:Integer} = For(f, θ)
 
-function For(f::F, θ::T) where {F, N, J <: Integer, T <: NTuple{N,J}}
+function For(f::F, θ::T) where {F,N,J<:Integer,T<:NTuple{N,J}}
     d = f.(Ones{Int}(N)...)
     D = typeof(d)
     X = eltype(d)
-    For{F, NTuple{N,J}, D, X}(f,θ)
+    return For{F,NTuple{N,J},D,X}(f, θ)
 end
 
-@inline function logpdf(d::For{F,T,D,X1},xs::AbstractArray{X2,N}) where {F, N, J <: Integer, T <: NTuple{N,J}, D,  X1,  X2}
+@inline function logpdf(
+    d::For{F,T,D,X1},
+    xs::AbstractArray{X2,N},
+) where {F,N,J<:Integer,T<:NTuple{N,J},D,X1,X2}
     s = 0.0
     @inbounds @simd for θ in CartesianIndices(d.θ)
         s += logpdf(d.f(Tuple(θ)...), xs[θ])
     end
-    s
+    return s
 end
 
-function Base.rand(dist::For) 
-    map(CartesianIndices(dist.θ)) do I
+function Base.rand(dist::For)
+    return map(CartesianIndices(dist.θ)) do I
         (rand ∘ dist.f)(Tuple(I)...)
     end
 end
@@ -44,27 +47,28 @@ end
 # T <: NTuple{N,J} where {J <: AbstractUnitRange}
 #########################################################
 
-For(f, θ::J...) where {J <: AbstractUnitRange} = For(f,θ)
+For(f, θ::J...) where {J<:AbstractUnitRange} = For(f, θ)
 
-function For(f::F, θ::T) where {F, N, J <: AbstractRange, T <: NTuple{N,J}}
+function For(f::F, θ::T) where {F,N,J<:AbstractRange,T<:NTuple{N,J}}
     d = f.(ones(Int, N)...)
     D = typeof(d)
     X = eltype(d)
-    For{F, NTuple{N,J}, D, X}(f,θ)
+    return For{F,NTuple{N,J},D,X}(f, θ)
 end
 
-
-@inline function logpdf(d::For{F,T,D,X1},xs::AbstractArray{X2,N}) where {F, N, J <: AbstractRange,  T <: NTuple{N,J}, D, X1, X2}
+@inline function logpdf(
+    d::For{F,T,D,X1},
+    xs::AbstractArray{X2,N},
+) where {F,N,J<:AbstractRange,T<:NTuple{N,J},D,X1,X2}
     s = 0.0
     @inbounds @simd for θ in CartesianIndices(d.θ)
         s += logpdf(d.f(Tuple(θ)...), xs[θ])
     end
-    s
+    return s
 end
 
-
-function Base.rand(dist::For{F,T}) where {F,  N, J <: AbstractRange, T <: NTuple{N,J}}
-    map(CartesianIndices(dist.θ)) do I
+function Base.rand(dist::For{F,T}) where {F,N,J<:AbstractRange,T<:NTuple{N,J}}
+    return map(CartesianIndices(dist.θ)) do I
         (rand ∘ dist.f)(Tuple(I)...)
     end
 end
@@ -116,7 +120,7 @@ end
 end
 
[…]
