### 2020-07-11 dc9994a
CompatHelper: bump compat for "Setfield" to "0.7"
