+++ Project.toml
@@ -58,7 +58,7 @@ RecipesBase = "0.7,0.8, 1.0"
 Reexport = "0.2"
 ResumableFunctions = "0.5"
 ReverseDiff = "1.0"
-Setfield = "0.6"
+Setfield = "0.6, 0.7"
 SimpleGraphs = "0.3"
 SimplePartitions = "0.2"
 SimplePosets = "0.0"
