### 2020-07-13 b6c765a
CompatHelper: bump compat for "SimpleGraphs" to "0.4" (#154)
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Chad Scherrer <chad.scherrer@gmail.com>
