### 2020-07-13 26cb180
CompatHelper: bump compat for "SimplePartitions" to "0.3" (#152)
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Chad Scherrer <chad.scherrer@gmail.com>
