+++ Project.toml
@@ -61,7 +61,8 @@ ResumableFunctions = "0.5"
 ReverseDiff = "1.0"
 SimpleGraphs = "0.3, 0.4"
 Setfield = "0.6, 0.7"
-SimplePartitions = "0.2"
+SimpleGraphs = "0.3"
+SimplePartitions = "0.2, 0.3"
 SimplePosets = "0.0, 0.1"
 SpecialFunctions = "0.10"
 StatsFuns = "0.9"
