+++ Project.toml
@@ -59,9 +59,8 @@ Reexport = "0.2"
 Requires = "1.0"
 ResumableFunctions = "0.5"
 ReverseDiff = "1.0"
-SimpleGraphs = "0.3, 0.4"
 Setfield = "0.6, 0.7"
-SimpleGraphs = "0.3"
+SimpleGraphs = "0.3, 0.4"
 SimplePartitions = "0.2, 0.3"
 SimplePosets = "0.0, 0.1"
 SpecialFunctions = "0.10"
