### 2020-07-18 430631e
CompatHelper: bump compat for "SimpleGraphs" to "0.5"
