### 2020-07-18 6991b3e
CompatHelper: bump compat for "SimpleGraphs" to "0.5"
