### 2020-07-20 8b042e7
Fixed merge and transforms so they handle return values. Updated Tests for this
