### 2020-07-20 b5013be
removed outdated comment
