+++ src/core/model.jl
@@ -53,8 +53,7 @@ function Base.merge(m1::Model, m2::Model)
     vals = merge(m1.vals, m2.vals)
     args = setdiff(union(m1.args, m2.args), keys(vals))
     dists = merge(m1.dists, m2.dists)
-    retn = merge_retn(m2.retn, m1.retn) # m2 first so it gets priority
-
+    retn = merge_retn(m1.retn, m2.retn)
     Model(theModule, args, vals, dists, retn)
 end
 
