+++ Project.toml
@@ -42,7 +42,7 @@ AdvancedHMC = "0.2"
 DiffResults = "1.0"
 Distributions = "0.21, 0.22, 0.23"
 DynamicHMC = "2.1"
-FillArrays = "0.8"
+FillArrays = "0.8, 0.9"
 ForwardDiff = "0.10"
 GeneralizedGenerated = "0.2"
 Graphs = "0.10"
