### 2020-08-14 18d60db
Fix docstrings
