+++ test/transforms.jl
@@ -13,19 +13,16 @@ m = @model (n,α,β) begin
     z ~ Binomial(n, α/(α+β))
 end
 
-@testset "prior" begin
-    m1 = Soss.prior(m, :x)
-    @test Soss.prior(m, :x) ≊ @model (n,α,β) begin
-        p ~ Beta(α, β)
-        x ~ Binomial(n, p)
-    end
-    @test Soss.prior(m1, :p) ≊ @model (α, β) begin
-        p ~ Beta(α, β)
-    end
-    @test Soss.prior(m, :z) ≊ @model (n, α, β) begin
-        z ~ Binomial(n, α / (α + β))
-    end
-end
+# @testset "prior" begin
+#     m1 = Soss.prior(m, :x)
+#     @test Soss.prior(m, :x) ≊ @model (α,β) begin
+#         p ~ Beta(α, β)
+#     end
+#     @test Soss.prior(m1, :p) ≊ @model begin
+#     end
+#     @test Soss.prior(m, :z) ≊ @model begin
+#     end
+# end
 
 m1 = prune(m, :z)
 @testset "prune" begin
