+++ .travis.yml
@@ -30,7 +30,7 @@ jobs:
       julia: 1.4
       os: linux
       script:
-        - julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd()));
-                                               Pkg.instantiate()'
-        - julia --project=docs/ docs/make.jl
+        - PYTHON=conda julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd()));
+                                                            Pkg.instantiate()'
+        - PYTHON=conda julia --project=docs/ docs/make.jl
       after_success: skip
