### 2020-08-14 36b2567
Make sure that we also run Travis CI on the latest stable release of Julia
Also, fix the Travis CI documentation build
