+++ .travis.yml
@@ -25,8 +25,8 @@ after_success:
 - julia -e 'using Pkg; Pkg.add("Coverage"); using Coverage; Coveralls.submit(process_folder())'
 
 jobs:
-allow_failures:
-  - julia: 'nightly'
+  allow_failures:
+    - julia: 'nightly'
   include:
     - stage: "Documentation"
       julia: '1.4'
