### 2020-08-16 fb78235
Remove the dependency on ResumableFunctions.jl
