+++ Project.toml
@@ -25,7 +25,6 @@ Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 RecipesBase = "3cdcf5f2-1ef4-517c-9805-6587b60abb01"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 Requires = "ae029012-a4dd-5104-9daa-d747884805df"
-ResumableFunctions = "c5292f4c-5179-55e1-98c5-05642aab7184"
 ReverseDiff = "37e2e3b7-166d-5795-8a7a-e32c996b4267"
 Setfield = "efcf1570-3423-57d1-acb7-fd33fddbac46"
 SimpleGraphs = "55797a34-41de-5266-9ec1-32ac4eb504d3"
