### 2020-08-16 bf396fc
Turn off deprecation warnings during the test suite
