### 2020-08-16 7bf6381
Manually forward the `depwarn` flag to the test process
