### 2020-08-17 ac6f0a0
Remove ResumableFunctions dependency
