### 2020-08-17 74ebbad
Comment out dynamicHMC method requiring ResumableFunctions
