### 2020-08-17 1d11bfd
update doctests
