### 2020-08-17 6fee418
update doctests
