+++ src/inference/advancedhmc.jl
@@ -51,7 +51,7 @@ println("Posterior mean β: " * string(round(E_β, digits=2)))
 
 # output
 true β: -0.3
-Posterior mean β: -0.25
+Posterior mean β: -0.26
 ```
 
 
