+++ Project.toml
@@ -1,7 +1,7 @@
 name = "Soss"
 uuid = "8ce77f84-9b61-11e8-39ff-d17a774bf41c"
 author = ["Chad Scherrer <chad.scherrer@gmail.com>"]
-version = "0.13.0"
+version = "0.13.1"
 
 [deps]
 AdvancedHMC = "0bf59076-c3b1-5ca4-86bd-e02cd72cde3d"
@@ -43,7 +43,7 @@ Distributions = "0.21, 0.22, 0.23"
 DynamicHMC = "2.1"
 FillArrays = "0.8"
 ForwardDiff = "0.10"
-GeneralizedGenerated = "0.2"
+GeneralizedGenerated = "0.2.6"
 Graphs = "0.10"
 IterTools = "1.3"
 LazyArrays = "0.14, 0.15, 0.16"
