+++ README.md
@@ -36,6 +36,7 @@ and start again. More information on julia projects [here](https://julialang.git
 Let's jump right in with a simple linear model:
 
 ````julia
+
 using Soss
 
 m = @model X begin
@@ -65,7 +66,7 @@ julia> import Random; Random.seed!(3)
 Random.MersenneTwister(UInt32[0x00000003], Random.DSFMT.DSFMT_state(Int32[-1359582567, 1073454075, 1934390716, 1073583786, -114685834, 1073112842, -1913218479, 1073122729, -73577195, 1073266439  …  1226759590, 1072980451, -1366384707, 1073012992, 1661148031, 2121090155, 141576524, -658637225, 382, 0]), [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0  …  0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0], UInt128[0x00000000000000000000000000000000, 0x00000000000000000000000000000000, 0x00000000000000000000000000000000, 0x00000000000000000000000000000000, 0x00000000000000000000000000000000, 0x00000000000000000000000000000000, 0x00000000000000000000000000000000, 0x00000000000000000000000000000000, 0x00000000000000000000000000000000, 0x00000000000000000000000000000000  …  0x00000000000000000000000000000000, 0x00000000000000000000000000000000, 0x00000000000000000000000000000000, 0x00000000000000000000000000000000, 0x00000000000000000000000000000000, 0x00000000000000000000000000000000, 0x00000000000000000000000000000000, 0x00000000000000000000000000000000, 0x00000000000000000000000000000000, 0x00000000000000000000000000000000], 1002, 0)
 
 julia> X = randn(6,2)
-6×2 Array{Float64,2}:
+6×2 Matrix{Float64}:
   1.19156    0.100793
  -2.51973   -0.00197414
   2.07481    1.00879
@@ -79,12 +80,12 @@ julia> X = randn(6,2)
 
 ````julia
 julia> truth = rand(m(X=X));
-(β = [0.07187269298745927, -0.5128103336795292], y = [0.10079289135480324, -2.5197330871745263, 2.0748097755419757, 0.8442227439533416, 1.158074626662026, -0.47515878362112707])
+(β = [0.07187269298745927, -0.5128103336795292], y = [-1.5462858466884173, 0.33157118320250245, -1.07546820508531, -1.6689735918627429, -1.0335681260821046, -0.5761553015966487])
 
 julia> pairs(truth)
 pairs(::NamedTuple) with 2 entries:
   :β => [0.0718727, -0.51281]
-  :y => [0.100793, -2.51973, 2.07481, 0.844223, 1.15807, -0.475159]
+  :y => [-1.54629, 0.331571, -1.07547, -1.66897, -1.03357, -0.576155]
 
 ````
 
@@ -92,7 +93,7 @@ pairs(::NamedTuple) with 2 entries:
 
 ````julia
 julia> truth.β
-2-element Array{Float64,1}:
+2-element Vector{Float64}:
   0.07187269298745927
  -0.5128103336795292
 
@@ -102,13 +103,13 @@ julia> truth.β
 
 ````julia
 julia> truth.y
-6-element Array{Float64,1}:
-  0.10079289135480324
- -2.5197330871745263
-  2.0748097755419757
-  0.8442227439533416
-  1.158074626662026
- -0.47515878362112707
+6-element Vector{Float64}:
+ -1.5462858466884173
+  0.33157118320250245
+ -1.07546820508531
+ -1.6689735918627429
+ -1.0335681260821046
+ -0.5761553015966487
 
 ````
 
[…]
