### 2020-08-31 d499c26
add `statements` function, simplify `withdistributions`
