+++ src/core/statement.jl
@@ -53,3 +53,6 @@ function Base.convert(::Type{Expr}, sts::Vector{Statement})
     Expr(:block, [convert(Expr, st) for st in sts]...)
 end
 
+function statements(m::Model)
+    Statement[Soss.findStatement(m, v) for v in variables(m)]
+end
