### 2020-08-31 ebd5f33
add `statements` function, simplify `withdistributions`
