+++ src/transforms/withdistributions.jl
@@ -57,7 +57,6 @@ function withdistributions(m::Model)
     proc(st) = Model(theModule, st)
 
     # Rewrite the statements of the model one by one. 
-    # This would be a bit cleaner with a `statements` iterator
     m_new = foldl(statements(m); init=m_init) do m0,st
         merge(m0, proc(st))
     end
