### 2020-08-31 bb6fad5
Update toposort.jl
Tests are failing since SimpleGraphs 0.5.2. Changing to see if tests pass now. Will we need to also update compat?
