### 2020-08-31 cd72544
`weave()` no longer takes "throw_errors" kwarg. It's a chunk-option now.
