+++ test/readme.jl
@@ -2,7 +2,8 @@ using Weave
 using Documenter
 
 function buildREADME()
-    weave("../README.jmd", doctype= "github", throw_errors=true, cache=:refresh, args=Dict(:seed => 6))
+    set_chunk_defaults!(:error=>false)
+    weave("../README.jmd", doctype= "github", cache=:refresh, args=Dict(:seed => 6))
     return true
 end
 
