### 2020-08-31 91f0d7e
add RNG argument
