### 2020-09-01 8ab1dd8
CompatHelper: add new compat entry for "CategoricalArrays" at version "0.8"
