### 2020-09-01 1821821
CompatHelper: add new compat entry for "Bijectors" at version "0.8"
