### 2020-09-02 6e38700
Add rand(::AbstractRNG, ::MarkovChain) method
