### 2020-09-02 932bdef
rand(::AbstractRNG, ::For)
