### 2020-09-02 588d68d
More `rand` methods for `For`
