+++ src/distributions/for.jl
@@ -70,11 +70,13 @@ end
     return s
 end
 
-function Base.rand(dist::For{F,T}) where {F,N,J<:AbstractRange,T<:NTuple{N,J}}
+function Base.rand(rng::AbstractRNG, dist::For{F,T}) where {F,N,J<:AbstractRange,T<:NTuple{N,J}}
     return map(CartesianIndices(dist.θ)) do I
-        (rand ∘ dist.f)(Tuple(I)...)
+        rand(rng, dist.f(Tuple(I)...))
     end
 end
+
+Base.rand(dist::For{F,T}) where {F,N,J<:AbstractRange,T<:NTuple{N,J}} = rand(GLOBAL_RNG, dist)
 
 #########################################################
 # T <: Base.Generator
@@ -95,9 +97,11 @@ end
     return s
 end
 
-@inline function rand(d::For{F,T,D,X}) where {F,T<:Base.Generator,D,X}
-    return rand.(Base.Generator(d.f ∘ d.θ.f, d.θ.iter))
+@inline function Base.rand(rng::AbstractRNG, d::For{F,T,D,X}) where {F,T<:Base.Generator,D,X}
+    return rand.(rng, Base.Generator(d.f ∘ d.θ.f, d.θ.iter))
 end
+
+Base.rand(d::For{F,T,D,X}) where {F,T<:Base.Generator,D,X} = rand(GLOBAL_RNG, d)
 
 #########################################################
 # T <: AbstractArray
@@ -118,9 +122,11 @@ end
     return s
 end
 
-@inline function rand(d::For{F,T,D,X}) where {F,T<:AbstractArray,D,X}
+@inline function Base.rand(rng::AbstractRNG, d::For{F,T,D,X}) where {F,T<:AbstractArray,D,X}
     return rand.(d.f.(d.θ))
 end
+
+Base.rand(d::For{F,T,D,X}) where {F,T<:AbstractArray,D,X} = rand(GLOBAL_RNG, d)
 
 export logpdf2
 @inline function logpdf2(d::For{F,N,X1}, xs) where {F,N,X1,X2}
