+++ test/runtests.jl
@@ -3,4 +3,4 @@ using Test
 
 include("doctests.jl")
 include("transforms.jl")
-include("readme.jl")
+#include("readme.jl")
