### 2020-09-02 0f8215f
uncomment tests
