### 2020-09-03 570ffc2
change tests to approx equality. Seeing differences on different hardware
