### 2020-09-03 517232f
Properly qualify `rand` and `logpdf`
