+++ src/symbolic/codegen.jl
@@ -157,6 +157,6 @@ end
 # end
 
 
-function logpdf(m::JointDistribution{A0,A,B,M},x,::typeof(codegen)) where {A0,A,B,M}
+function Distributions.logpdf(m::JointDistribution{A0,A,B,M},x,::typeof(codegen)) where {A0,A,B,M}
     codegen(M, m.model, m.args, x)
 end
