### 2020-09-03 58c4bf0
Update src/distributions/for.jl
