+++ src/distributions/for.jl
@@ -97,7 +97,6 @@ end
     return s
 end
 
-
 @inline function Base.rand(rng::AbstractRNG, d::For{F,T,D,X}) where {F,T<:Base.Generator,D,X}
     return rand.(rng, Base.Generator(d.f ∘ d.θ.f, d.θ.iter))
 end
