### 2020-09-03 e0fee6b
Properly qualify `rand` and `logpdf` (#183)
* Properly qualify `rand` and `logpdf`
