+++ src/distributions/for.jl
@@ -27,7 +27,7 @@ function For(f::F, θ::T) where {F,N,J<:Integer,T<:NTuple{N,J}}
     return For{F,NTuple{N,J},D,X}(f, θ)
 end
 
-@inline function logpdf(
+@inline function Distributions.logpdf(
     d::For{F,T,D,X1},
     xs::AbstractArray{X2,N},
 ) where {F,N,J<:Integer,T<:NTuple{N,J},D,X1,X2}
@@ -59,7 +59,7 @@ function For(f::F, θ::T) where {F,N,J<:AbstractRange,T<:NTuple{N,J}}
     return For{F,NTuple{N,J},D,X}(f, θ)
 end
 
-@inline function logpdf(
+@inline function Distributions.logpdf(
     d::For{F,T,D,X1},
     xs::AbstractArray{X2,N},
 ) where {F,N,J<:AbstractRange,T<:NTuple{N,J},D,X1,X2}
@@ -89,7 +89,7 @@ function For(f::F, θ::T) where {F,T<:Base.Generator}
     return For{F,T,D,X}(f, θ)
 end
 
-@inline function logpdf(d::For{F,T}, x) where {F,T<:Base.Generator}
+@inline function Distributions.logpdf(d::For{F,T}, x) where {F,T<:Base.Generator}
     s = 0.0
     for (θj, xj) in zip(d.θ, x)
         s += logpdf(d.f(θj), xj)
@@ -114,7 +114,7 @@ function For(f::F, θ::T) where {F,T<:AbstractArray}
     return For{F,T,D,X}(f, θ)
 end
 
-@inline function logpdf(d::For{F,T}, x) where {F,T<:AbstractArray}
+@inline function Distributions.logpdf(d::For{F,T}, x) where {F,T<:AbstractArray}
     s = 0.0
     @inbounds @simd for j in eachindex(d.θ)
         s += logpdf(d.f(d.θ[j]), x[j])
@@ -128,7 +128,6 @@ end
 
 Base.rand(d::For{F,T,D,X}) where {F,T<:AbstractArray,D,X} = rand(GLOBAL_RNG, d)
 
-export logpdf2
 @inline function logpdf2(d::For{F,N,X1}, xs) where {F,N,X1,X2}
     results = zeros(eltype(xs), nthreads())
 
