### 2020-09-04 6eb0fd9
Omnibus PR: GitHub Actions CI, fix tests, reorganize docs, other improvements
