+++ docs/src/api.md
@@ -1,0 +1,12 @@
+```@meta
+CurrentModule = Soss
+```
+
+# API
+
+```@index
+```
+
+```@autodocs
+Modules = [Soss]
+```
