+++ docs/src/index.md
@@ -1,8 +1,10 @@
-# Soss.jl
-
-```@index
+```@meta
+CurrentModule = Soss
 ```
 
-```@autodocs
-Modules = [Soss]
-```
+# Soss
+
+[Soss](https://github.com/cscherrer/Soss.jl)
+is a library for _probabilistic programming_.
+
+The source code is available in the [GitHub repository](https://github.com/cscherrer/Soss.jl).
