+++ CITATION.bib
@@ -2,7 +2,6 @@
 	author  = {Chad Scherrer},
 	title   = {{Soss.jl}},
 	url     = {https://github.com/cscherrer/Soss.jl},
-	version = {v0.5.0},
 	year    = {2019},
 	month   = {9}
 }
