+++ previews/PR144/siteinfo.js
@@ -1,1 +1,0 @@
-var DOCUMENTER_CURRENT_VERSION = "previews/PR144";
