### 2020-09-04 3a42de7
Remove the PyCall/SymPy stuff
