### 2020-09-04 f60861f
Fix doctests
