+++ Project.toml
@@ -65,7 +65,7 @@ SimplePosets = "0.0, 0.1"
 SpecialFunctions = "0.10"
 StatsFuns = "0.9"
 TransformVariables = "0.3"
-julia = "1.3"
+julia = "1.4"
 
 [extras]
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
