### 2020-09-04 a2dc949
Omnibus PR: GitHub Actions CI, fix tests, reorganize docs, remove PyCall/SymPy stuff (#184)
