+++ test/doctests.jl
@@ -1,9 +1,12 @@
 using Documenter
+using Random
+using Soss
+using StableRNGs
 
 DocMeta.setdocmeta!(Soss, :DocTestSetup,
     quote
-        using Soss
         using Random
+        using Soss
         using StableRNGs
         Random.seed!(3)
     end; recursive=true)
