+++ readme/build_readme.jl
@@ -1,4 +1,0 @@
-using Weave
-
-# Run from Soss.jl/
-weave("readme/README.jmd", out_path="./", doctype= "github", cache=:refresh, args=Dict(:seed => 6))

[…]
