### 2020-09-04 8339019
Add and update multiple `[compat]` entries (#181)
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
