+++ Project.toml
@@ -39,16 +39,18 @@ TransformVariables = "84d833dd-6860-57f9-a1a7-6da5db126cff"
 
 [compat]
 AdvancedHMC = "0.2"
+CategoricalArrays = "0.8"
+Bijectors = "0.8"
 DiffResults = "1.0"
 Distributions = "0.21, 0.22, 0.23"
 DistributionsAD = "0.6"
 DynamicHMC = "2.1"
-FillArrays = "0.8"
+FillArrays = "0.8, 0.9"
 ForwardDiff = "0.10"
 GeneralizedGenerated = "0.2"
 Graphs = "0.10"
 IterTools = "1.3"
-LazyArrays = "0.14, 0.15, 0.16"
+LazyArrays = "0.14, 0.15, 0.16, 0.17"
 LogDensityProblems = "0.10"
 MLStyle = "0.3, 0.4"
 MacroTools = "0.5"
