### 2020-09-04 a2a54bb
Bump version number from "0.14.0-DEV" to "0.14.0" (#197)
