### 2020-09-04 a0935ab
Tweaks to the linear regression example (#198)
