+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"api/","page":"API","title":"API","text":"CurrentModule = Soss","category":"page"},{"location":"api/#API","page":"API","title":"API","text":"","category":"section"},{"location":"api/","page":"API","title":"API","text":"","category":"page"},{"location":"api/","page":"API","title":"API","text":"Modules = [Soss]","category":"page"},{"location":"api/#Soss.MarkovChain","page":"API","title":"Soss.MarkovChain","text":"MarkovChain\n\nMarkovChain(pars, step) defines a Markov Chain with global parameters pars and transition kernel step. Here, pars is a named tuple, and step is a Soss model that takes arguments (pars, state) and returns a next value containing the new pars and state.\n\nNOTE: This is experimental, and may change in the near future.\n\nmstep = @model pars,state begin\n    σ = pars.σ\n    x0 = state.x\n    x ~ Normal(x0, σ)\n    next = (pars=pars, state=(x=x,))\nend;\n\nm = @model s0 begin\n    σ ~ Exponential()\n    pars = (σ=σ,)\n    chain ~ MarkovChain(pars, mstep(pars=pars, state=s0))\nend;\n\nr = rand(m(s0=(x=2,),));\n\nfor s in Iterators.take(r.chain,3)\n    println(s)\nend\n\n# output\n\n(x = -6.596883394256064,)\n(x = 0.48200039561318864,)\n(x = -2.838556784903994,)\n\n\n\n\n\n","category":"type"},{"location":"api/#Soss.Do-Tuple{Model,Vararg{Any,N} where N}","page":"API","title":"Soss.Do","text":"Do(m, xs...)\n\nReturns a model transformed by adding xs... to arguments. The remainder of the body remains the same, consistent with Judea Pearl's \"Do\" operator. Unneeded arguments are trimmed.\n\nExamples\n\nm = @model (n, k) begin\n    β ~ Gamma()\n    α ~ Gamma()\n    θ ~ Beta(α, β)\n    x ~ Binomial(n, θ)\n    z ~ Binomial(k, α / (α + β))\nend;\nDo(m, :θ)\n\n# output\n@model (n, k, θ) begin\n        β ~ Gamma()\n        α ~ Gamma()\n        x ~ Binomial(n, θ)\n        z ~ Binomial(k, α / (α + β))\n    end\n\n\n\n\n\n","category":"method"},{"location":"api/#Soss.advancedHMC-Union{Tuple{B}, Tuple{A}, Tuple{Soss.JointDistribution{A,B,B1,M} where M where B1,Any}, Tuple{Soss.JointDistribution{A,B,B1,M} where M where B1,Any,Any}} where B where A","page":"API","title":"Soss.advancedHMC","text":"advancedHMC(m, data, N = 1000; n_adapts = 1000)\n\nDraw N samples from the posterior distribution of parameters defined in Soss model m, conditional on data. Samples are drawn using Hamiltonial Monte Carlo (HMC) from the advancedHMC.jl package.\n\nKeywords\n\nn_adapts = 1000: The number of interations used to set HMC parameters.\n\nReturns a tuple of length 2:\n\nSamples from the posterior distribution of parameters.\nSample summary statistics.\n\nExample\n\nm = @model x begin\n    β ~ Normal()\n    yhat = β .* x\n    y ~ For(eachindex(x)) do j\n        Normal(yhat[j], 2.0)\n    end\nend\n\nx = randn(10);\ntruth = [0.61, -0.34, -1.74];\n\npost = advancedHMC(m(x=x), (y=truth,));\nE_β = mean(post[1])[1]\n\nprintln(\"Posterior mean β: \" * string(round(E_β, digits=2
[…]
