### 2020-09-04 3553498
Docs: Add more comments to the linear regression example
