### 2020-09-04 5f0c786
Docs: Add more comments to the linear regression example (#199)
* Docs: Add more comments to the linear regression example

* Fix several comments

* Bump version from "0.14.1" to "0.14.2"
