+++ docs/src/sossmlj.md
@@ -1,0 +1,12 @@
+```@meta
+CurrentModule = Soss
+```
+
+# SossMLJ.jl
+
+The
+[SossMLJ.jl](https://github.com/cscherrer/SossMLJ.jl)
+package integrates Soss.jl into the MLJ.jl machine learning framework.
+
+More details are available in the
+[SossMLJ.jl GitHub repository](https://github.com/cscherrer/SossMLJ.jl).
