### 2020-09-06 b19134b
Fix rand(::EqualMix)
