### 2020-09-07 f96455a
CompatHelper: bump compat for "LazyArrays" to "0.18" (#202)
* CompatHelper: bump compat for "LazyArrays" to "0.18"

* Update Project.toml

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Dilum Aluthge <dilum@aluthge.com>
