+++ Project.toml
@@ -1,7 +1,7 @@
 name = "Soss"
 uuid = "8ce77f84-9b61-11e8-39ff-d17a774bf41c"
 author = ["Chad Scherrer <chad.scherrer@gmail.com>"]
-version = "0.14.3"
+version = "0.14.4"
 
 [deps]
 AdvancedHMC = "0bf59076-c3b1-5ca4-86bd-e02cd72cde3d"
@@ -39,8 +39,8 @@ TransformVariables = "84d833dd-6860-57f9-a1a7-6da5db126cff"
 
 [compat]
 AdvancedHMC = "0.2"
-CategoricalArrays = "0.8"
 Bijectors = "0.8"
+CategoricalArrays = "0.8"
 DiffResults = "1.0"
 Distributions = "0.21, 0.22, 0.23"
 DistributionsAD = "0.6"
@@ -50,7 +50,7 @@ ForwardDiff = "0.10"
 GeneralizedGenerated = "0.2"
 Graphs = "0.10"
 IterTools = "1.3"
-LazyArrays = "0.14, 0.15, 0.16, 0.17"
+LazyArrays = "0.14, 0.15, 0.16, 0.17, 0.18"
 LogDensityProblems = "0.10"
 MLStyle = "0.3, 0.4"
 MacroTools = "0.5"
