### 2020-09-07 09743be
more methods for `For`
