### 2020-09-09 a534c50
more methods for `For` (#204)
* more methods for `For`
* add tests

Co-authored-by: Dilum Aluthge <dilum@aluthge.com>
