### 2020-09-10 a4d0ec6
Revert "more methods for `For` (#204)"
This reverts commit a534c5081a59a49a70441d8a57269e01fb964078.
