### 2020-09-10 7137fc0
Revert "more methods for `For` (#204)" (#208)
This reverts commit a534c5081a59a49a70441d8a57269e01fb964078.
