### 2020-09-10 6f1c19b
Bump version number to 0.15.0
