### 2020-09-10 d2d764b
More methods for `For` - Take 2 (#209)
* more methods for `For`

Co-authored-by: Dilum Aluthge <dilum@aluthge.com>
