### 2020-09-14 6de0db7
CompatHelper: bump compat for "SimpleGraphs" to "0.6"
