### 2020-09-14 6a68dfe
CompatHelper: bump compat for "SimpleGraphs" to "0.6" (#210)
* CompatHelper: bump compat for "SimpleGraphs" to "0.6"

* Update Project.toml

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Dilum Aluthge <dilum@aluthge.com>
