+++ Project.toml
@@ -1,7 +1,7 @@
 name = "Soss"
 uuid = "8ce77f84-9b61-11e8-39ff-d17a774bf41c"
 author = ["Chad Scherrer <chad.scherrer@gmail.com>"]
-version = "0.15.0"
+version = "0.15.1"
 
 [deps]
 AdvancedHMC = "0bf59076-c3b1-5ca4-86bd-e02cd72cde3d"
@@ -61,7 +61,7 @@ Reexport = "0.2"
 Requires = "1.0"
 ReverseDiff = "1.0"
 Setfield = "0.6, 0.7"
-SimpleGraphs = "0.3, 0.4, 0.5"
+SimpleGraphs = "0.3, 0.4, 0.5, 0.6"
 SimplePartitions = "0.2, 0.3"
 SimplePosets = "0.0, 0.1"
 SpecialFunctions = "0.10"
