### 2020-09-29 edc0f98
CompatHelper: bump compat for "SimpleGraphs" to "0.7"
