+++ Project.toml
@@ -61,7 +61,7 @@ Reexport = "0.2"
 Requires = "1.0"
 ReverseDiff = "1.0"
 Setfield = "0.6, 0.7"
-SimpleGraphs = "0.3, 0.4, 0.5, 0.6"
+SimpleGraphs = "0.3, 0.4, 0.5, 0.6, 0.7"
 SimplePartitions = "0.2, 0.3"
 SimplePosets = "0.0, 0.1"
 SpecialFunctions = "0.10"
