### 2020-09-29 5aec7eb
CompatHelper: bump compat for "SimpleGraphs" to "0.7" (#212)
* CompatHelper: bump compat for "SimpleGraphs" to "0.7"

* Update Project.toml

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Dilum Aluthge <dilum@aluthge.com>
