+++ Project.toml
@@ -42,7 +42,7 @@ AdvancedHMC = "0.2"
 Bijectors = "0.8"
 CategoricalArrays = "0.8"
 DiffResults = "1.0"
-Distributions = "0.21, 0.22, 0.23"
+Distributions = "0.21, 0.22, 0.23, 0.24"
 DistributionsAD = "0.6"
 DynamicHMC = "2.1"
 FillArrays = "0.8, 0.9"
