### 2020-10-10 39f1767
CompatHelper: bump compat for "Distributions" to "0.24" (#213)
* CompatHelper: bump compat for "Distributions" to "0.24"

* Bump version

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Chad Scherrer <chad.scherrer@gmail.com>
