+++ Project.toml
@@ -1,7 +1,7 @@
 name = "Soss"
 uuid = "8ce77f84-9b61-11e8-39ff-d17a774bf41c"
 author = ["Chad Scherrer <chad.scherrer@gmail.com>"]
-version = "0.15.2"
+version = "0.15.3"
 
 [deps]
 AdvancedHMC = "0bf59076-c3b1-5ca4-86bd-e02cd72cde3d"
@@ -42,7 +42,7 @@ AdvancedHMC = "0.2"
 Bijectors = "0.8"
 CategoricalArrays = "0.8"
 DiffResults = "1.0"
-Distributions = "0.21, 0.22, 0.23"
+Distributions = "0.21, 0.22, 0.23, 0.24"
 DistributionsAD = "0.6"
 DynamicHMC = "2.1"
 FillArrays = "0.8, 0.9"
