+++ src/importance.jl
@@ -9,25 +9,31 @@ end
 @gg M function _importanceSample(_::Type{M}, p::Model, _pargs, q::Model, _qargs, _data) where M <: TypeLevel{Module}
     p = type2model(p)
     q = type2model(q)
-
+        
     Expr(:let,
         Expr(:(=), :M, from_type(M)),
-        sourceImportanceSample()(p,q) |> loadvals(_qargs, _data) |> loadvals(_pargs, NamedTuple()))
+        sourceImportanceSample(_data)(p,q) |> loadvals(_qargs, _data) |> loadvals(_pargs, NamedTuple()) |> merge_pqargs)
+
+
 end
 
-sourceImportanceSample(p::Model,q::Model) = sourceImportanceSample()(p::Model,q::Model)
+sourceImportanceSample(p::Model,q::Model,_data) = sourceImportanceSample(_data)(p::Model,q::Model)
 
 export sourceImportanceSample
-function sourceImportanceSample()
+function sourceImportanceSample(_data)
     function(p::Model,q::Model)
         p = canonical(p)
         q = canonical(q)
         m = merge(p,q)
 
+        _datakeys = getntkeys(_data)
+
         function proc(m, st::Sample) 
+            st.x ∈ _datakeys && return :(_ℓ += logpdf($(st.rhs), $(st.x)))
+
             if hasproperty(p.dists, st.x)
                 pdist = getproperty(p.dists, st.x)
-                qdist = st.rhs
+                qdist = getproperty(q.dists, st.x)
                 @gensym ℓx
                 result = @q begin
                     $ℓx = importanceSample($pdist, $qdist, _data)
@@ -35,7 +41,7 @@ function sourceImportanceSample()
                     $(st.x) = $ℓx.val
                 end
                 return flatten(result)
-            else return :($(st.x) = $(st.rhs))
+            else return :($(st.x) = rand($(st.rhs)))
             end
             return flatten(result)
         end
@@ -45,11 +51,11 @@ function sourceImportanceSample()
 
         body = buildSource(m, proc) |> flatten
 
-        kwargs = freeVariables(q) ∪ arguments(p)
+        kwargs = arguments(p) ∪ arguments(q)
         kwargsExpr = Expr(:tuple,kwargs...)
 
         stochExpr = begin
-            vals = map(sampled(m)) do x Expr(:(=), x,x) end
+            vals = map(sampled(q)) do x Expr(:(=), x,x) end
             Expr(:tuple, vals...)
         end
 
@@ -64,7 +70,7 @@ function sourceImportanceSample()
 end
 
 @inline function importanceSample(p, q, _data)
-    x = merge(rand(q), _data)
+    x = rand(q)
     ℓ = logpdf(p,x) - logpdf(q,x)
     Weighted(ℓ,x)
 end
@@ -139,3 +145,10 @@ end
 #     ))
 
 # end
+
+function merge_pqargs(src)
+    @q begin
+        _args = merge(_pargs, _qargs)
+        $src
+    end |> flatten
+end
