### 2020-10-21 63218e9
CompatHelper: add new compat entry for "MappedArrays" at version "0.3"
