+++ Project.toml
@@ -55,6 +55,7 @@ LazyArrays = "0.14, 0.15, 0.16, 0.17, 0.18"
 LogDensityProblems = "0.10"
 MLStyle = "0.3, 0.4"
 MacroTools = "0.5"
+MappedArrays = "0.3"
 MonteCarloMeasurements = "0.6,0.7, 0.8, 0.9"
 NamedTupleTools = "0.12, 0.13"
 RecipesBase = "0.7,0.8, 1.0"
