+++ src/Soss.jl
@@ -16,6 +16,7 @@ using LazyArrays
 using FillArrays
 using Requires
 
+include("core/abstractmodel.jl")
 include("core/statement.jl")
 include("core/model.jl")
 include("core/jointdistribution.jl")
@@ -24,6 +25,7 @@ include("core/dependencies.jl")
 include("core/toposort.jl")
 include("core/weighted.jl")
 include("core/utils.jl")
+include("core/conditional.jl")
 
 include("distributions/dist.jl")
 include("distributions/for.jl")
