### 2020-10-23 e6744f5
CompatHelper: add new compat entry for "MappedArrays" at version "0.3" (#215)
* CompatHelper: add new compat entry for "MappedArrays" at version "0.3"

* version bump

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Chad Scherrer <chad.scherrer@gmail.com>
