+++ Project.toml
@@ -46,7 +46,7 @@ DiffResults = "1.0"
 Distributions = "0.21, 0.22, 0.23, 0.24"
 DistributionsAD = "0.6"
 DynamicHMC = "2.1"
-FillArrays = "0.8, 0.9"
+FillArrays = "0.8, 0.9, 0.10"
 ForwardDiff = "0.10"
 GeneralizedGenerated = "0.2"
 Graphs = "0.10"
