### 2020-10-26 885bd0f
CompatHelper: bump compat for "LazyArrays" to "0.19"
