+++ Project.toml
@@ -51,7 +51,7 @@ ForwardDiff = "0.10"
 GeneralizedGenerated = "0.2"
 Graphs = "0.10"
 IterTools = "1.3"
-LazyArrays = "0.14, 0.15, 0.16, 0.17, 0.18"
+LazyArrays = "0.14, 0.15, 0.16, 0.17, 0.18, 0.19"
 LogDensityProblems = "0.10"
 MLStyle = "0.3, 0.4"
 MacroTools = "0.5"
