+++ Project.toml
@@ -1,7 +1,7 @@
 name = "Soss"
 uuid = "8ce77f84-9b61-11e8-39ff-d17a774bf41c"
 author = ["Chad Scherrer <chad.scherrer@gmail.com>"]
-version = "0.15.5"
+version = "0.15.6"
 
 [deps]
 AdvancedHMC = "0bf59076-c3b1-5ca4-86bd-e02cd72cde3d"
@@ -46,7 +46,7 @@ DiffResults = "1.0"
 Distributions = "0.21, 0.22, 0.23, 0.24"
 DistributionsAD = "0.6"
 DynamicHMC = "2.1"
-FillArrays = "0.8, 0.9"
+FillArrays = "0.8, 0.9, 0.10"
 ForwardDiff = "0.10"
 GeneralizedGenerated = "0.2"
 Graphs = "0.10"
