### 2020-10-26 #2062 @cscherrer
Thanks @arfon . I think resubmission would be fine. We have some changes to our SymPy approach that we should update anyway. This was my first JOSS submission with the current system; next time I'll work through the checklist before hitting submit. Thanks to everyone here for your time.
